% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acen_hg19}
\alias{acen_hg19}
\title{centromere regions (hg19)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 3 columns.
}
\usage{
acen_hg19
}
\description{
centromere regions (hg19)
}
\keyword{datasets}
