% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acen_hg38}
\alias{acen_hg38}
\title{centromere regions (hg38)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 3 columns.
}
\usage{
acen_hg38
}
\description{
centromere regions (hg38)
}
\keyword{datasets}
