% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_counts}
\alias{aggregate_counts}
\title{Utility function to make reference gene expression profiles}
\usage{
aggregate_counts(count_mat, cell_annot, verbose = T)
}
\arguments{
\item{count_mat}{raw count matrices where rownames are genes and column names are cells}

\item{cell_annot}{dataframe with columns "cell" and "cell_type"}

\item{verbose}{verbosity}
}
\value{
a matrix of reference gene expression levels
}
\description{
Utility function to make reference gene expression profiles
}
