% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annot_cm}
\alias{annot_cm}
\title{Annotate genetic distance between markers}
\usage{
annot_cm(bulk, genetic_map)
}
\arguments{
\item{bulk}{pseudobulk dataframe}

\item{genetic_map}{a genetic map dataframe}
}
\value{
a pseudobulk dataframe
}
\description{
Annotate genetic distance between markers
}
\keyword{internal}
