% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annot_consensus}
\alias{annot_consensus}
\title{Annotate a consensus segments on a pseudobulk dataframe}
\usage{
annot_consensus(bulk, segs_consensus)
}
\arguments{
\item{bulk}{a pseudobulk dataframe}

\item{segs_consensus}{a consensus segment dataframe}
}
\value{
a pseudobulk dataframe
}
\description{
Annotate a consensus segments on a pseudobulk dataframe
}
