% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annot_ref}
\alias{annot_ref}
\title{example reference cell annotation}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\usage{
annot_ref
}
\description{
example reference cell annotation
}
\keyword{datasets}
