% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{annot_superclones}
\alias{annot_superclones}
\title{Annotate superclones on a tree}
\usage{
annot_superclones(gtree, geno, p_min = 0.95, precision_cutoff = 0.9)
}
\description{
Annotate superclones on a tree
}
