% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annot_theta_mle}
\alias{annot_theta_mle}
\title{Annotate the theta parameter for each segment}
\usage{
annot_theta_mle(bulk)
}
\arguments{
\item{bulk}{dataframe Pseudobulk profile}
}
\value{
dataframe Pseudobulk profile
}
\description{
Annotate the theta parameter for each segment
}
\keyword{internal}
