% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annot_theta_roll}
\alias{annot_theta_roll}
\title{Annotate rolling estimate of imbalance level theta}
\usage{
annot_theta_roll(bulk)
}
\arguments{
\item{bulk}{a pseudobulk dataframe}
}
\value{
a pseudobulk dataframe
}
\description{
Annotate rolling estimate of imbalance level theta
}
\keyword{internal}
