% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{annotate_genes}
\alias{annotate_genes}
\title{Annotate genes on allele dataframe}
\usage{
annotate_genes(df, gtf)
}
\arguments{
\item{df}{dataframe Allele count dataframe}

\item{gtf}{dataframe Gene gtf}
}
\value{
dataframe Allele dataframe with gene column
}
\description{
Annotate genes on allele dataframe
}
