% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{approx_phi_post}
\alias{approx_phi_post}
\title{Laplace approximation of the posterior of expression fold change phi}
\usage{
approx_phi_post(
  Y_obs,
  lambda_ref,
  d,
  alpha = NULL,
  beta = NULL,
  mu = NULL,
  sig = NULL,
  lower = 0.2,
  upper = 10,
  start = 1
)
}
\description{
Laplace approximation of the posterior of expression fold change phi
}
\keyword{internal}
