% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{approx_theta_post}
\alias{approx_theta_post}
\title{Laplace approximation of the posterior of allelic imbalance theta}
\usage{
approx_theta_post(
  pAD,
  DP,
  p_s,
  lower = 0.001,
  upper = 0.499,
  start = 0.25,
  gamma = 20
)
}
\description{
Laplace approximation of the posterior of allelic imbalance theta
}
\keyword{internal}
