% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_example}
\alias{bulk_example}
\title{example pseudobulk dataframe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3935 rows and 83 columns.
}
\usage{
bulk_example
}
\description{
example pseudobulk dataframe
}
\keyword{datasets}
