% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{calc_allele_lik}
\alias{calc_allele_lik}
\title{Calculate allele likelihoods}
\usage{
calc_allele_lik(pAD, DP, p_s, theta, gamma = 20)
}
\arguments{
\item{pAD}{integer vector Paternal allele counts}

\item{DP}{integer vector Total alelle counts}

\item{p_s}{numeric vector Phase switch probabilities}

\item{theta}{numeric Haplotype imbalance}

\item{gamma}{numeric Overdispersion in the allele-specific expression}
}
\description{
Calculate allele likelihoods
}
\keyword{internal}
