% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_allele_LLR}
\alias{calc_allele_LLR}
\title{Calculate LLR for an allele HMM}
\usage{
calc_allele_LLR(pAD, DP, p_s, theta_mle, theta_0 = 0, gamma = 20)
}
\arguments{
\item{pAD}{vector of phased allele depth}

\item{DP}{vector of total allele depth}

\item{p_s}{vector of phase switch probabilities}

\item{theta_mle}{MLE of imbalance level theta (alternative hypothesis)}

\item{theta_0}{imbalance level in the null hypothesis}

\item{gamma}{dispersion parameter for the Beta-Binomial allele model}
}
\value{
a log-likelihood ratio
}
\description{
Calculate LLR for an allele HMM
}
\keyword{internal}
