% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_cluster_dist}
\alias{calc_cluster_dist}
\title{Calculate expression distance matrix between cell populatoins}
\usage{
calc_cluster_dist(count_mat, cell_annot)
}
\arguments{
\item{count_mat}{dgCMatrix Gene expression counts}

\item{cell_annot}{dataframe specifying the cell ID and cluster memberships}
}
\value{
a distance matrix
}
\description{
Calculate expression distance matrix between cell populatoins
}
\keyword{internal}
