% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_exp_LLR}
\alias{calc_exp_LLR}
\title{Calculate LLR for an expression HMM}
\usage{
calc_exp_LLR(
  Y_obs,
  lambda_ref,
  d,
  phi_mle,
  mu = NULL,
  sig = NULL,
  alpha = NULL,
  beta = NULL,
  model = "lnpois"
)
}
\arguments{
\item{Y_obs}{vector of gene expression counts}

\item{lambda_ref}{vector of reference expression levels}

\item{d}{vector of library sizes}

\item{phi_mle}{MLE of expression fold change phi (alternative hypothesis)}

\item{mu}{mean parameter for the PLN expression model}

\item{sig}{dispersion parameter for the PLN expression model}

\item{alpha}{hyperparameter for the gamma poisson model (not used)}

\item{beta}{hyperparameter for the gamma poisson model (not used)}
}
\value{
a log-likelihood ratio
}
\description{
Calculate LLR for an expression HMM
}
\keyword{internal}
