% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_phi_mle_lnpois}
\alias{calc_phi_mle_lnpois}
\title{Calculate the MLE of expression fold change phi}
\usage{
calc_phi_mle_lnpois(Y_obs, lambda_ref, d, mu, sig, lower = 0.1, upper = 10)
}
\description{
Calculate the MLE of expression fold change phi
}
\keyword{internal}
