% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cell_to_clone}
\alias{cell_to_clone}
\title{Map cells to the phylogeny (or genotypes) based on CNV posteriors}
\usage{
cell_to_clone(gtree, exp_post, allele_post)
}
\arguments{
\item{gtree}{a cell lineage tree as a tidygraph object}

\item{exp_post}{expression posteriors}

\item{allele_post}{allele posteriors}
}
\value{
clone posteriors
}
\description{
Map cells to the phylogeny (or genotypes) based on CNV posteriors
}
