% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_allele_df}
\alias{check_allele_df}
\title{Check the format of a allele dataframe}
\usage{
check_allele_df(df)
}
\arguments{
\item{df}{dataframe Allele dataframe}
}
\value{
dataframe Allele dataframe
}
\description{
Check the format of a allele dataframe
}
\keyword{internal}
