% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_contam}
\alias{check_contam}
\title{check inter-individual contamination}
\usage{
check_contam(bulk)
}
\arguments{
\item{bulk}{dataframe Pseudobulk profile}
}
\description{
check inter-individual contamination
}
\keyword{internal}
