% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_exp_noise}
\alias{check_exp_noise}
\title{check noise level}
\usage{
check_exp_noise(bulk)
}
\arguments{
\item{bulk}{dataframe Pseudobulk profile}
}
\description{
check noise level
}
\keyword{internal}
