% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_exp_ref}
\alias{check_exp_ref}
\title{check the format of lambdas_ref}
\usage{
check_exp_ref(lambdas_ref)
}
\arguments{
\item{lambdas_ref}{matrix Expression reference profile}
}
\value{
matrix Expression reference profile
}
\description{
check the format of lambdas_ref
}
\keyword{internal}
