% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_matrix}
\alias{check_matrix}
\title{Check the format of a count matrix}
\usage{
check_matrix(count_mat)
}
\arguments{
\item{count_mat}{matrix Count matrix}
}
\value{
matrix Count matrix
}
\description{
Check the format of a count matrix
}
\keyword{internal}
