% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_segs_fix}
\alias{check_segs_fix}
\title{check the format of a given consensus segment dataframe}
\usage{
check_segs_fix(segs_consensus_fix)
}
\arguments{
\item{segs_consensus_fix}{dataframe Consensus segment dataframe}
}
\value{
dataframe Consensus segment dataframe
}
\description{
check the format of a given consensus segment dataframe
}
\keyword{internal}
