% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_segs_loh}
\alias{check_segs_loh}
\title{Check the format of a given clonal LOH segment dataframe}
\usage{
check_segs_loh(segs_loh)
}
\arguments{
\item{segs_loh}{dataframe Clonal LOH segment dataframe}
}
\value{
dataframe Clonal LOH segment dataframe
}
\description{
Check the format of a given clonal LOH segment dataframe
}
\keyword{internal}
