% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_ref_cor}
\alias{choose_ref_cor}
\title{choose beest reference for each cell based on correlation}
\usage{
choose_ref_cor(count_mat, lambdas_ref, gtf_transcript)
}
\arguments{
\item{count_mat}{gene expression count matrix}

\item{lambdas_ref}{reference expression matrix}

\item{gtf_transcript}{transcript info dataframe}
}
\value{
best references for each cell
}
\description{
choose beest reference for each cell based on correlation
}
