% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrom_sizes_hg19}
\alias{chrom_sizes_hg19}
\title{chromosome sizes (hg19)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 22 rows and 2 columns.
}
\usage{
chrom_sizes_hg19
}
\description{
chromosome sizes (hg19)
}
\keyword{datasets}
