% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrom_sizes_hg38}
\alias{chrom_sizes_hg38}
\title{chromosome sizes (hg38)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 22 rows and 2 columns.
}
\usage{
chrom_sizes_hg38
}
\description{
chromosome sizes (hg38)
}
\keyword{datasets}
