% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_alleles}
\alias{classify_alleles}
\title{classify alleles using viterbi and forward-backward}
\usage{
classify_alleles(bulk)
}
\arguments{
\item{bulk}{a pesudobulk dataframe}
}
\value{
a pesudobulk dataframe
}
\description{
classify alleles using viterbi and forward-backward
}
