% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{cnv_heatmap}
\alias{cnv_heatmap}
\title{Plot CNV heatmap}
\usage{
cnv_heatmap(
  segs,
  var = "group",
  label_group = TRUE,
  legend = TRUE,
  exclude_gap = TRUE
)
}
\arguments{
\item{segs}{dataframe Segments to plot. Need columns "seg_start", "seg_end", "cnv_state"}

\item{var}{character Column to facet by}

\item{label_group}{logical Label the groups}

\item{legend}{logical Display the legend}
}
\description{
Plot CNV heatmap
}
