% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{contract_nodes}
\alias{contract_nodes}
\title{Merge adjacent set of nodes}
\usage{
contract_nodes(G, vset, node_tar = NULL, debug = FALSE)
}
\arguments{
\item{G}{igraph Mutation graph}

\item{vset}{vector Set of adjacent vertices to merge}
}
\value{
igraph Mutation graph
}
\description{
Merge adjacent set of nodes
}
\keyword{internal}
