% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_mat_ATC2}
\alias{count_mat_ATC2}
\title{something}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 33538 rows and 173 columns.
}
\usage{
count_mat_ATC2
}
\description{
something
}
\keyword{datasets}
