% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_mat_example}
\alias{count_mat_example}
\title{example gene expression count matrix}
\format{
An object of class \code{dgCMatrix} with 1024 rows and 173 columns.
}
\usage{
count_mat_example
}
\description{
example gene expression count matrix
}
\keyword{datasets}
