% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_mat_ref}
\alias{count_mat_ref}
\title{example reference count matrix}
\format{
An object of class \code{dgCMatrix} with 1000 rows and 50 columns.
}
\usage{
count_mat_ref
}
\description{
example reference count matrix
}
\keyword{datasets}
