% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{dbbinom}
\alias{dbbinom}
\title{Beta-binomial distribution density function
A distribution is beta-binomial if p, the probability of success, 
in a binomial distribution has a beta distribution with shape 
parameters alpha > 0 and beta > 0
For more details, see extraDistr::dbbinom}
\usage{
dbbinom(x, size, alpha = 1, beta = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{size}{number of trials (zero or more)}

\item{alpha}{numeric (default=1)}

\item{beta}{numeric (default=1)}

\item{log}{boolean (default=FALSE)}
}
\value{
density values returned as numeric vector
}
\description{
Beta-binomial distribution density function
A distribution is beta-binomial if p, the probability of success, 
in a binomial distribution has a beta distribution with shape 
parameters alpha > 0 and beta > 0
For more details, see extraDistr::dbbinom
}
\keyword{internal}
