% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_clonal_loh}
\alias{detect_clonal_loh}
\title{Call clonal LOH using SNP density. 
Rcommended for cell lines or tumor samples with no normal cells.}
\usage{
detect_clonal_loh(bulk, t = 1e-05, min_depth = 0)
}
\arguments{
\item{bulk}{dataframe Pseudobulk profile}

\item{t}{numeric Transition probability}

\item{min_depth}{integer Minimum coverage to filter SNPs}
}
\value{
dataframe LOH segments
}
\description{
Call clonal LOH using SNP density. 
Rcommended for cell lines or tumor samples with no normal cells.
}
\examples{
bulk = get_bulk(
   count_mat,
   ref_hca,
   df_allele,
   gtf_hg38,
   genetic_map_hg38
)
segs_loh = bulk \%>\% detect_loh(t = 1e-5)
}
