% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_allele_ATC2}
\alias{df_allele_ATC2}
\title{something}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 600845 rows and 25 columns.
}
\usage{
df_allele_ATC2
}
\description{
something
}
\keyword{datasets}
