% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_allele_example}
\alias{df_allele_example}
\title{example allele count dataframe}
\format{
An object of class \code{data.frame} with 41167 rows and 11 columns.
}
\usage{
df_allele_example
}
\description{
example allele count dataframe
}
\keyword{datasets}
