% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dgpois}
\alias{dgpois}
\title{Density function for a gamma-poisson distribution}
\usage{
dgpois(x, shape, rate, scale = 1/rate, log = FALSE)
}
\description{
Density function for a gamma-poisson distribution
}
\keyword{internal}
