% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poilog.R
\name{dpoilog}
\alias{dpoilog}
\title{Returns the density for the Poisson lognormal distribution with parameters mu and sig}
\usage{
dpoilog(x, mu, sig, log = FALSE)
}
\arguments{
\item{x}{vector of integers, the observations}

\item{mu}{mean of lognormal distribution}

\item{sig}{standard deviation of lognormal distribution}

\item{log}{boolean Return the log density if TRUE (default=FALSE)}
}
\description{
Returns the density for the Poisson lognormal distribution with parameters mu and sig
}
\keyword{internal}
