% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{exp_hclust}
\alias{exp_hclust}
\title{Run smoothed expression-based hclust}
\usage{
exp_hclust(
  count_mat_obs,
  lambdas_ref,
  gtf_transcript,
  multi_ref = F,
  k = 5,
  ncores = 10,
  verbose = T
)
}
\description{
Run smoothed expression-based hclust
}
