% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{expand_states}
\alias{expand_states}
\title{expand multi-allelic CNVs into separate entries in the single-cell posterior dataframe}
\usage{
expand_states(sc_post, segs_consensus)
}
\description{
expand multi-allelic CNVs into separate entries in the single-cell posterior dataframe
}
\keyword{internal}
