% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_genes}
\alias{filter_genes}
\title{filter for mutually expressed genes}
\usage{
filter_genes(count_mat, lambdas_ref, gtf_transcript, verbose = T)
}
\arguments{
\item{count_mat}{observed gene count matrices}

\item{lambdas_ref}{expression values in reference profile}

\item{gtf_transcript}{transcript dataframe}
}
\value{
list of genes that are kept after filtering
}
\description{
filter for mutually expressed genes
}
\keyword{internal}
