% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_common_diploid}
\alias{find_common_diploid}
\title{Find the common diploid region in a group of pseudobulks}
\usage{
find_common_diploid(
  bulks,
  grouping = "clique",
  gamma = 20,
  theta_min = 0.08,
  t = 1e-05,
  fc_min = 1.25,
  alpha = 1e-04,
  ncores = 5,
  debug = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bulks}{combined dataframe from mutiple pseudobulks (differentiated by "sample" column)}

\item{grouping}{whether to use cliques or components in the graph to find dipoid cluster}

\item{gamma}{dispersion parameter for beta-binomial allele model}

\item{theta_min}{minimum imbalance threshold}

\item{t}{transition probability for the allele HMM}

\item{fc_min}{minimum fold change to call quadruploid cluster}

\item{alpha}{FDR cut-off for q values to determine edges}

\item{ncores}{number of cores to use}
}
\value{
a list of info
}
\description{
Find the common diploid region in a group of pseudobulks
}
