% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_gamma}
\alias{fit_gamma}
\title{fit gamma maximum likelihood}
\usage{
fit_gamma(AD, DP, start = 20)
}
\arguments{
\item{AD}{numeric vector Variant allele depth}

\item{DP}{numeric vector Total allele depth}
}
\value{
a fit
}
\description{
fit gamma maximum likelihood
}
\keyword{internal}
