% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_lnpois}
\alias{fit_lnpois}
\title{fit a PLN model by maximum likelihood}
\usage{
fit_lnpois(Y_obs, lambda_ref, d)
}
\arguments{
\item{Y_obs}{gene expression counts}

\item{lambda_ref}{reference expression levels}

\item{d}{total library size}
}
\value{
MLE of mu and sig
}
\description{
fit a PLN model by maximum likelihood
}
\keyword{internal}
