% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_multi_ref}
\alias{fit_multi_ref}
\title{fit reference weights for a pseudobulk expression profile}
\usage{
fit_multi_ref(
  Y_obs,
  lambdas_ref,
  d,
  gtf_transcript,
  min_lambda = 2e-06,
  verbose = FALSE
)
}
\arguments{
\item{Y_obs}{gene expression count vector}

\item{lambdas_ref}{reference expression matrix}

\item{d}{total library size}

\item{gtf_transcript}{transcript info dataframe}

\item{min_lambda}{expression level threshold to filter genes}
}
\value{
fitted reference weights and expression model parameters
}
\description{
fit reference weights for a pseudobulk expression profile
}
\keyword{internal}
