% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{forward_back_allele}
\alias{forward_back_allele}
\title{Forward-backward algorithm for allele HMM}
\usage{
forward_back_allele(obj, ...)
}
\description{
Forward-backward algorithm for allele HMM
}
\keyword{internal}
