% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaps_hg19}
\alias{gaps_hg19}
\title{genome gap regions (hg19)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 28 rows and 3 columns.
}
\usage{
gaps_hg19
}
\description{
genome gap regions (hg19)
}
\keyword{datasets}
