% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaps_hg38}
\alias{gaps_hg38}
\title{genome gap regions (hg38)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 30 rows and 3 columns.
}
\usage{
gaps_hg38
}
\description{
genome gap regions (hg38)
}
\keyword{datasets}
