% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_postfix}
\alias{generate_postfix}
\title{Generate alphabetical postfixes}
\usage{
generate_postfix(n)
}
\arguments{
\item{n}{vector of integers}
}
\value{
vector of alphabetical postfixes
}
\description{
Generate alphabetical postfixes
}
\keyword{internal}
