% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genetic_map_hg38}
\alias{genetic_map_hg38}
\title{something}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3281217 rows and 6 columns.
}
\usage{
genetic_map_hg38
}
\description{
something
}
\keyword{datasets}
