% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.r
\name{genotype}
\alias{genotype}
\title{Main function}
\usage{
genotype(label, samples, vcfs, outdir)
}
\arguments{
\item{label}{character string}

\item{samples}{list of strings}

\item{vcfs}{vcfR objects}
}
\value{
a status code
}
\description{
Main function
}
