% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_allele_bulk}
\alias{get_allele_bulk}
\title{Aggregate into pseudobulk alelle profile}
\usage{
get_allele_bulk(df_allele, genetic_map, lambda = 1, min_depth = 0)
}
\arguments{
\item{df_allele}{single cell allele dataframe}

\item{genetic_map}{a genetic map dataframe}

\item{lambda}{phase switch rate}

\item{min_depth}{minimum coverage to filter SNPs by}
}
\value{
a dataframe of bulk allele profile
}
\description{
Aggregate into pseudobulk alelle profile
}
\keyword{internal}
