% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_allele_post}
\alias{get_allele_post}
\title{get CNV allele posteriors}
\usage{
get_allele_post(bulks, segs_consensus, df_allele, naive = FALSE)
}
\arguments{
\item{bulks}{dataframe Subtree pseudobulk profiles}

\item{segs_consensus}{dataframe Consensus CNV segments}

\item{df_allele}{dataframe Allele counts}
}
\value{
dataframe Allele posteriors
}
\description{
get CNV allele posteriors
}
\keyword{internal}
