% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_clone_post}
\alias{get_clone_post}
\title{Map cells to the phylogeny (or genotypes) based on CNV posteriors}
\usage{
get_clone_post(gtree, exp_post, allele_post)
}
\arguments{
\item{gtree}{tbl_graph A cell lineage tree}

\item{exp_post}{dataframe Expression posteriors}

\item{allele_post}{dataframe Allele posteriors}
}
\value{
dataframe Clone posteriors
}
\description{
Map cells to the phylogeny (or genotypes) based on CNV posteriors
}
\keyword{internal}
