% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_exp_bulk}
\alias{get_exp_bulk}
\title{Aggregate into bulk expression profile}
\usage{
get_exp_bulk(count_mat, lambdas_ref, gtf_transcript, verbose = TRUE)
}
\arguments{
\item{count_mat}{observed gene count matrices}

\item{lambdas_ref}{expression values in reference profile}

\item{gtf_transcript}{transcript dataframe}
}
\value{
a dataframe of bulk gene expression profile
}
\description{
Aggregate into bulk expression profile
}
\keyword{internal}
