% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_exp_likelihoods_lnpois}
\alias{get_exp_likelihoods_lnpois}
\title{get the single cell expression likelihoods}
\usage{
get_exp_likelihoods_lnpois(
  exp_sc,
  diploid_chroms = NULL,
  use_loh = FALSE,
  depth_obs = NULL,
  mu = NULL,
  sigma = NULL
)
}
\arguments{
\item{exp_sc}{dataframe Single-cell expression counts}
}
\description{
get the single cell expression likelihoods
}
\keyword{internal}
