% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_exp_post}
\alias{get_exp_post}
\title{compute single-cell expression posteriors}
\usage{
get_exp_post(
  segs_consensus,
  count_mat,
  gtf_transcript,
  lambdas_ref,
  sc_refs = NULL,
  diploid_chroms = NULL,
  use_loh = NULL,
  ncores = 30,
  verbose = TRUE,
  debug = F
)
}
\arguments{
\item{segs_consensus}{consensus segments}

\item{count_mat}{gene expression count matrix}

\item{gtf_transcript}{transcript dataframe}

\item{lambdas_ref}{reference expression matrix}
}
\value{
expression posterior datatframe
}
\description{
compute single-cell expression posteriors
}
\keyword{internal}
