% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_exp_sc}
\alias{get_exp_sc}
\title{get the single cell expression dataframe}
\usage{
get_exp_sc(segs_consensus, count_mat, gtf_transcript)
}
\arguments{
\item{segs_consensus}{consensus segments}

\item{count_mat}{gene expression count matrix}

\item{gtf_transcript}{transcript dataframe}
}
\value{
single cell expression dataframe
}
\description{
get the single cell expression dataframe
}
\keyword{internal}
