% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_gtree}
\alias{get_gtree}
\title{Get a tidygraph tree with simplified mutational history}
\usage{
get_gtree(tree, P, n_cut = 0, max_cost = 0)
}
\arguments{
\item{tree}{phylo Single-cell phylogenetic tree}

\item{P}{matrix Genotype probability matrix}

\item{n_cut}{integer Number of cuts on the phylogeny (resulting in n_cut + 1 clones)}

\item{max_cost}{numeric Likelihood threshold to collapse internal branches}
}
\value{
tbl_graph Phylogeny annotated with branch lengths and mutation events
}
\description{
Get a tidygraph tree with simplified mutational history
}
