% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_haplotype_post}
\alias{get_haplotype_post}
\title{Get phased haplotypes}
\usage{
get_haplotype_post(bulks, segs_consensus, naive = FALSE)
}
\arguments{
\item{bulks}{dataframe Subtree pseudobulk profiles}

\item{segs_consensus}{dataframe Consensus CNV segments}

\item{naive}{logical Whether to use naive haplotype classification}
}
\value{
dataframe Posterior haplotypes
}
\description{
Get phased haplotypes
}
\keyword{internal}
