% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inter_cm}
\alias{get_inter_cm}
\title{Helper function to get inter-SNP distance}
\usage{
get_inter_cm(d)
}
\arguments{
\item{d}{numeric vector Genetic positions in centimorgan (cM)}
}
\value{
numeric vector Inter-SNP genetic distances
}
\description{
Helper function to get inter-SNP distance
}
\keyword{internal}
