% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_internal_nodes}
\alias{get_internal_nodes}
\title{Helper function to get the internal nodes of a dendrogram and the leafs in each subtree}
\usage{
get_internal_nodes(den, node, labels)
}
\description{
Helper function to get the internal nodes of a dendrogram and the leafs in each subtree
}
\keyword{internal}
