% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_joint_post}
\alias{get_joint_post}
\title{get joint posteriors}
\usage{
get_joint_post(exp_post, allele_post, segs_consensus)
}
\arguments{
\item{exp_post}{dataframe Expression single-cell CNV posteriors}

\item{allele_post}{dataframe Allele single-cell CNV posteriors}

\item{segs_consensus}{dataframe Consensus CNV segments}
}
\value{
dataframe Joint single-cell CNV posteriors
}
\description{
get joint posteriors
}
\keyword{internal}
