% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_move_cost}
\alias{get_move_cost}
\title{Get the cost of a mutation reassignment}
\usage{
get_move_cost(muts, node_ori, node_tar, l_matrix)
}
\arguments{
\item{muts}{character Mutations dlimited by comma}

\item{node_ori}{character Name of the "from" node}

\item{node_tar}{character Name of the "to" node}
}
\value{
numeric Likelihood cost of the mutation reassignment
}
\description{
Get the cost of a mutation reassignment
}
\keyword{internal}
