% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_move_opt}
\alias{get_move_opt}
\title{Get the least costly mutation reassignment}
\usage{
get_move_opt(G, l_matrix)
}
\arguments{
\item{G}{igraph Mutation graph}

\item{l_matrix}{matrix Likelihood matrix of mutation placements}
}
\value{
numeric Lieklihood cost of performing the mutation move
}
\description{
Get the least costly mutation reassignment
}
\keyword{internal}
