% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_segs_consensus}
\alias{get_segs_consensus}
\title{Extract consensus CNV segments}
\usage{
get_segs_consensus(bulks, min_LLR = 20, min_overlap = 0.45)
}
\arguments{
\item{bulks}{pseudobulks dataframe}

\item{min_LLR}{LLR threshold to filter CNVs}

\item{min_overlap}{minimum overlap fraction to determine count two events as as overlapping}
}
\value{
consensus segments dataframe
}
\description{
Extract consensus CNV segments
}
