% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_segs_neu}
\alias{get_segs_neu}
\title{get neutral segments from multiple pseudobulks}
\usage{
get_segs_neu(bulks)
}
\arguments{
\item{bulks}{combined dataframe from mutiple pseudobulks (differentiated by "sample" column)}
}
\value{
dataframe of neutral segments
}
\description{
get neutral segments from multiple pseudobulks
}
\keyword{internal}
