% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{get_snps}
\alias{get_snps}
\title{process VCFs into SNP dataframe}
\usage{
get_snps(vcf)
}
\arguments{
\item{vcf}{vcfR object}
}
\value{
dataframe SNP information
}
\description{
process VCFs into SNP dataframe
}
\keyword{internal}
