% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{get_trans_probs}
\alias{get_trans_probs}
\title{Helper function to calculate transition porbabilities}
\usage{
get_trans_probs(t, p_s, w, cn_from, phase_from, cn_to, phase_to)
}
\description{
Helper function to calculate transition porbabilities
}
\keyword{internal}
