% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{get_tree_post}
\alias{get_tree_post}
\title{Find maximum lilkelihood assignment of mutations on a tree}
\usage{
get_tree_post(tree, P)
}
\arguments{
\item{tree}{phylo Single-cell phylogenetic tree}

\item{P}{matrix Genotype probability matrix}
}
\value{
list Mutation
}
\description{
Find maximum lilkelihood assignment of mutations on a tree
}
\keyword{internal}
