% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gexp_roll_example}
\alias{gexp_roll_example}
\title{example smoothed gene expression dataframe}
\format{
An object of class \code{data.frame} with 10 rows and 2000 columns.
}
\usage{
gexp_roll_example
}
\description{
example smoothed gene expression dataframe
}
\keyword{datasets}
