% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtf_hg38}
\alias{gtf_hg38}
\title{gene model (hg38)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 26807 rows and 5 columns.
}
\usage{
gtf_hg38
}
\description{
gene model (hg38)
}
\keyword{datasets}
