% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtf_mm10}
\alias{gtf_mm10}
\title{gene model (mm10)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 30336 rows and 5 columns.
}
\usage{
gtf_mm10
}
\description{
gene model (mm10)
}
\keyword{datasets}
