% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joint_post_example}
\alias{joint_post_example}
\title{example joint single-cell cnv posterior dataframe}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3806 rows and 71 columns.
}
\usage{
joint_post_example
}
\description{
example joint single-cell cnv posterior dataframe
}
\keyword{datasets}
