% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_bbinom}
\alias{l_bbinom}
\title{calculate joint likelihood of allele data}
\usage{
l_bbinom(AD, DP, alpha, beta)
}
\arguments{
\item{AD}{variant allele depth}

\item{DP}{total allele depth}

\item{alpha}{alpha parameter of Beta-Binomial distribution}

\item{beta}{beta parameter of Beta-Binomial distribution}
}
\value{
joint log likelihood
}
\description{
calculate joint likelihood of allele data
}
\keyword{internal}
