% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_gpois}
\alias{l_gpois}
\title{calculate joint likelihood of a gamma-poisson model}
\usage{
l_gpois(Y_obs, lambda_ref, d, alpha, beta, phi = 1)
}
\arguments{
\item{Y_obs}{gene expression counts}

\item{lambda_ref}{reference expression levels}

\item{d}{total library size}

\item{alpha}{alpha parameter of Gamma-Poisson distribution}

\item{beta}{beta parameter of Gamma-Poisson distribution}
}
\value{
joint log likelihood
}
\description{
calculate joint likelihood of a gamma-poisson model
}
\keyword{internal}
