% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{l_s_v}
\alias{l_s_v}
\title{Compute site branch likelihood (not used)}
\usage{
l_s_v(node, site, gtree, geno)
}
\arguments{
\item{node}{integer Node id}

\item{site}{character Mutation site name}

\item{gtree}{tbl_graph The single-cell phylogeny}

\item{geno}{matrix Genotype probability matrix}
}
\value{
numeric Likelihood of assigning the mutation to this node
}
\description{
Compute site branch likelihood (not used)
}
\keyword{internal}
