% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{label_edges}
\alias{label_edges}
\title{Annotate the direct upstream or downstream mutations on the edges}
\usage{
label_edges(G)
}
\arguments{
\item{G}{igraph Mutation graph}
}
\value{
igraph Mutation graph
}
\description{
Annotate the direct upstream or downstream mutations on the edges
}
\keyword{internal}
