% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{label_genotype}
\alias{label_genotype}
\title{Label the genotypes on a mutation graph}
\usage{
label_genotype(G)
}
\arguments{
\item{G}{igraph Mutation graph}
}
\value{
igraph Mutation graph
}
\description{
Label the genotypes on a mutation graph
}
\keyword{internal}
