% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{likelihood_allele}
\alias{likelihood_allele}
\title{Only compute total log likelihood from an allele HMM}
\usage{
likelihood_allele(hmm)
}
\arguments{
\item{hmm}{HMM object; expect variables x (allele depth), d (total depth),
logPi (log transition prob matrix), delta (prior for each state), 
alpha (alpha for each state), beta (beta for each state), 
states (states), p_s (phase switch probs)}
}
\description{
Only compute total log likelihood from an allele HMM
}
\keyword{internal}
