% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{log_message}
\alias{log_message}
\title{Log a message}
\usage{
log_message(msg, verbose = TRUE)
}
\arguments{
\item{msg}{string Message to log}

\item{verbose}{boolean Whether to print message to console}
}
\description{
Log a message
}
\keyword{internal}
