% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_psbulk}
\alias{make_psbulk}
\title{Utility function to make reference gene expression profiles}
\usage{
make_psbulk(count_mat, cell_annot, verbose = T)
}
\arguments{
\item{count_mat}{raw count matrices where rownames are genes and column names are cells}

\item{cell_annot}{dataframe with columns "cell" and "cell_type"}
}
\value{
a matrix of reference gene expression profiles
}
\description{
Utility function to make reference gene expression profiles
}
