% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{mark_tumor_lineage}
\alias{mark_tumor_lineage}
\title{mark the tumor lineage of a phylogeny}
\usage{
mark_tumor_lineage(gtree)
}
\arguments{
\item{gtree}{a tidygraph tree}
}
\value{
a tidygraph tree
}
\description{
mark the tumor lineage of a phylogeny
}
