% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_graph_example}
\alias{mut_graph_example}
\title{example mutation graph}
\format{
An object of class \code{igraph} of length 5.
}
\usage{
mut_graph_example
}
\description{
example mutation graph
}
\keyword{datasets}
