% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{mut_to_tree}
\alias{mut_to_tree}
\title{transfer mutation assignment onto a single-cell phylogeny}
\usage{
mut_to_tree(gtree, mut_nodes)
}
\arguments{
\item{gtree}{tbl_graph The single-cell phylogeny}

\item{mut_nodes}{dataframe Mutation placements}
}
\value{
tbl_graph A single-cell phylogeny with mutation placements
}
\description{
transfer mutation assignment onto a single-cell phylogeny
}
\keyword{internal}
