% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{nni}
\alias{nni}
\title{Generate tree neighbourhood}
\usage{
nni(tree, ncores = 1)
}
\arguments{
\item{tree}{phylo Single-cell phylogenetic tree}

\item{ncores}{integer Number of cores to use}
}
\description{
Generate tree neighbourhood
}
\keyword{internal}
