% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{nnin}
\alias{nnin}
\title{Perform the NNI at a specific branch}
\usage{
nnin(tree, n)
}
\arguments{
\item{tree}{phylo Single-cell phylogenetic tree}

\item{n}{integer Branch ID}
}
\description{
Perform the NNI at a specific branch
}
\keyword{internal}
