% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Numbat}
\alias{Numbat}
\title{numbat R6 class}
\description{
used to summarize results
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Numbat$new()}}
\item \href{#method-fetch_results}{\code{Numbat$fetch_results()}}
\item \href{#method-plot_phylo_heatmap}{\code{Numbat$plot_phylo_heatmap()}}
\item \href{#method-plot_exp_roll}{\code{Numbat$plot_exp_roll()}}
\item \href{#method-clone}{\code{Numbat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize numbat class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$new(out_dir, i = 2, gtf = gtf_hg38, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_dir}}{output directory}

\item{\code{i}}{get results from which iteration}

\item{\code{gtf}}{transcript gtf dataframe}

\item{\code{verbose}}{verbosity}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'numbat' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetch_results"></a>}}
\if{latex}{\out{\hypertarget{method-fetch_results}{}}}
\subsection{Method \code{fetch_results()}}{
fetch results from the numbat output directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$fetch_results(out_dir, i = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_dir}}{output directory}

\item{\code{i}}{get results from which iteration}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_phylo_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plot_phylo_heatmap}{}}}
\subsection{Method \code{plot_phylo_heatmap()}}{
plot the single-cell CNV calls in a heatmap and the corresponding phylogeny
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$plot_phylo_heatmap(
  annot = NULL,
  geno_bar = FALSE,
  pal_clone = NULL,
  p_min = 0.9,
  tip_length = 1,
  branch_width = 0.2,
  line_width = 0.1,
  tree_height = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{annot}}{named list of cell annotation}

\item{\code{geno_bar}}{where to show genotype annotation}

\item{\code{pal_clone}}{clone color palette}

\item{\code{p_min}}{minimum probability threshold to filter the calls}

\item{\code{tip_length}}{tree tip length}

\item{\code{branch_width}}{tree branch width}

\item{\code{line_width}}{heatmap line width}

\item{\code{tree_height}}{plotting height of the tree}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_exp_roll"></a>}}
\if{latex}{\out{\hypertarget{method-plot_exp_roll}{}}}
\subsection{Method \code{plot_exp_roll()}}{
plot window-smoothed expression profiles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$plot_exp_roll(k = 3, n_sample = 300)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{number of clusters}

\item{\code{n_sample}}{number of cells to subsample}

\item{\code{hc}}{where to show clone annotation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
