% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{numbat_subclone}
\alias{numbat_subclone}
\title{Main function to decompose tumor subclones}
\usage{
numbat_subclone(
  count_mat,
  lambdas_ref,
  df,
  gtf_transcript,
  genetic_map,
  cell_annot = NULL,
  out_dir = "./",
  t = 1e-05,
  init_method = "smooth",
  init_k = 3,
  sample_size = 450,
  min_cells = 200,
  max_cost = 150,
  max_iter = 2,
  min_depth = 0,
  ncores = 30,
  exp_model = "lnpois",
  gbuild = "hg38",
  verbose = TRUE
)
}
\arguments{
\item{count_mat}{raw count matrices where rownames are genes and column names are cells}

\item{lambdas_ref}{either a named vector with gene names as names and normalized expression as values, or a matrix where rownames are genes and columns are pseudobulk names}

\item{df}{dataframe of allele counts per cell, produced by preprocess_allele}

\item{gtf_transcript}{gtf dataframe of transcripts}
}
\value{
a status code
}
\description{
Main function to decompose tumor subclones
}
