% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{perform_nni}
\alias{perform_nni}
\title{maximum likelihood tree search via NNI}
\usage{
perform_nni(
  tree_init,
  P,
  max_iter = 100,
  eps = 0.01,
  ncores = 1,
  verbose = TRUE
)
}
\arguments{
\item{tree_init}{iintial tree as phylo object}

\item{P}{genotype probability matrix}

\item{max_iter}{maximum number of iterations}

\item{eps}{tolerance threshold in likelihood difference for stopping}

\item{ncores}{number of cores to use}

\item{verbose}{verbosity}
}
\value{
a list of trees corresponding to the rearrangement steps
}
\description{
maximum likelihood tree search via NNI
}
