% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phi_hat_seg}
\alias{phi_hat_seg}
\title{Estimate of expression fold change phi in a segment}
\usage{
phi_hat_seg(Y_obs, lambda_ref, d, mu, sig)
}
\arguments{
\item{Y_obs}{vector of expression counts}

\item{lambda_ref}{vector of reference expression levels}

\item{d}{total library size}

\item{mu}{mean of the PLN expression model}

\item{sig}{dispersion of the PLN expression model}
}
\value{
estimate of phi
}
\description{
Estimate of expression fold change phi in a segment
}
\keyword{internal}
