% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylogeny_example}
\alias{phylogeny_example}
\title{example single-cell phylogeny}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 345.
}
\usage{
phylogeny_example
}
\description{
example single-cell phylogeny
}
\keyword{datasets}
