% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_mut_history}
\alias{plot_mut_history}
\title{plot mutational history}
\usage{
plot_mut_history(
  G,
  clone_post = NULL,
  edge_label_size = 4,
  node_label_size = 6,
  node_size = 10,
  show_clone_size = TRUE,
  edge_label = TRUE,
  node_label = TRUE,
  horizontal = TRUE,
  pal = NULL
)
}
\arguments{
\item{G}{igraph Mutation history graph}

\item{clone_post}{dataframe Clone assignment posteriors}

\item{edge_label_size}{numeric Size of edge label (default=4)}

\item{node_label_size}{numeric Size of node label (default=6)}

\item{node_size}{numeric Size of nodes (default=10)}

\item{show_clone_size}{logical Whether to show clone size (default=TRUE)}

\item{edge_label}{logical Whether to label edges (default=TRUE)}

\item{node_label}{logical Whether to label nodes (default=TRUE)}

\item{horizontal}{logical Whether to use horizontal layout (default=TRUE)}

\item{pal}{named vector Node colors (default=NULL)}
}
\value{
ggplot object
}
\description{
plot mutational history
}
