% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_phylo_heatmap}
\alias{plot_phylo_heatmap}
\title{plot single-cell CNV calls along with the clonal phylogeny}
\usage{
plot_phylo_heatmap(
  gtree,
  joint_post,
  segs_consensus,
  annot = NULL,
  pal_annot = NULL,
  annot_title = "Annotation",
  annot_scale = NULL,
  clone_dict = NULL,
  clone_bar = FALSE,
  pal_clone = NULL,
  clone_title = "Genotype",
  clone_legend = TRUE,
  p_min = 0.9,
  line_width = 0.1,
  tree_height = 1,
  branch_width = 0.2,
  tip_length = 0.2,
  clone_line = FALSE,
  superclone = FALSE
)
}
\arguments{
\item{gtree}{tbl_graph The single-cell phylogeny}

\item{joint_post}{dataframe Joint single cell CNV posteriors}

\item{segs_consensus}{datatframe Consensus segment dataframe}

\item{annot}{named vector Cell annotations, keys are cell names (default=NULL)}

\item{pal_annot}{named vector Colors for cell annotations (default=NULL)}

\item{annot_title}{character Legend title for the annotation bar  (default=Annotation)}

\item{annot_scale}{ggplot scale Color scale for the annotation bar (default=NULL)}

\item{clone_dict}{named vector Clone annotations, mapping from cell name to clones (default=NULL)}

\item{clone_bar}{logical Whether to display clone bar plot (default=FALSE)}

\item{clone_title}{character Legend title for the clone bar (default=Genotype)}

\item{clone_legend}{logical Whether to display the clone legend (default=TRUE)}

\item{p_min}{numeric Probability threshold to display CNV calls (default=0.9)}

\item{line_width}{numeric Line width for CNV heatmap (default=0.1)}

\item{tree_height}{numeric Relative height of the phylogeny plot (default=1)}

\item{branch_width}{numeric Line width in the phylogeny (default=0.2)}

\item{tip_length}{numeric Length of tips in the phylogeny (default=0.2)}

\item{clone_line}{logical Whether to display borders for clones in the heatmap (default=FALSE)}

\item{superclone}{logical Wehether to display superclones in the clone bar (default=FALSE)}
}
\value{
ggplot panel
}
\description{
plot single-cell CNV calls along with the clonal phylogeny
}
