% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_sc_roll}
\alias{plot_sc_roll}
\title{Plot single-cell smoothed expression magnitude heatmap}
\usage{
plot_sc_roll(
  gexp_roll_wide,
  hc,
  k,
  gtf,
  lim = 0.8,
  n_sample = 300,
  reverse = TRUE,
  plot_tree = TRUE
)
}
\arguments{
\item{gexp_roll_wide}{matrix Cell x gene smoothed expression magnitudes}

\item{hc}{hclust Hierarchical clustring result}

\item{k}{integer Number of clusters (default=3)}

\item{gtf}{dataframe Transcript GTF}

\item{lim}{numeric Limit for expression magnitudes (default=0.8)}

\item{n_sample}{integer Number of cells to subsample (default=300)}

\item{reverse}{logical Whether to reverse the cell order (default=TRUE)}

\item{plot_tree}{logical Whether to plot the dendrogram (default=TRUE)}
}
\description{
Plot single-cell smoothed expression magnitude heatmap
}
