% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_sc_tree}
\alias{plot_sc_tree}
\title{Plot single-cell smoothed expression magnitude heatmap}
\usage{
plot_sc_tree(
  gtree,
  label_mut = TRUE,
  label_size = 3,
  dot_size = 2,
  branch_width = 0.5,
  tip = TRUE,
  tip_length = 0.5,
  pal_clone = NULL
)
}
\arguments{
\item{gtree}{tbl_graph The single-cell phylogeny}

\item{label_mut}{logical Whether to label mutations}

\item{label_size}{numeric Size of mutation labels}

\item{dot_size}{numeric Size of mutation nodes}

\item{branch_width}{numeric Width of branches in tree}

\item{tip}{logical Whether to plot tip point}

\item{tip_length}{numeric Length of the tips}

\item{pal_clone}{named vector Clone colors}
}
\value{
ggplot A single-cell phylogeny with mutation history labeled
}
\description{
Plot single-cell smoothed expression magnitude heatmap
}
\examples{
p = plot_sc_tree(phylogeny_example)
}
