% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pnorm.range.log}
\alias{pnorm.range.log}
\title{Get the total probability from a region of a normal pdf}
\usage{
pnorm.range.log(lower, upper, mu, sd)
}
\description{
Get the total probability from a region of a normal pdf
}
\keyword{internal}
