% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{preprocess_allele}
\alias{preprocess_allele}
\title{Preprocess allele data}
\usage{
preprocess_allele(sample, vcf_pu, vcf_phased, AD, DP, barcodes, gtf_transcript)
}
\arguments{
\item{sample}{sample label}

\item{vcf_pu}{pileup VCF from cell-snp-lite}

\item{vcf_phased}{phased VCF from eagle2}

\item{AD}{alt allele depth matrix from pileup}

\item{DP}{total allele depth matrix from pileup}

\item{barcodes}{list of barcodes from pileup}

\item{gtf_transcript}{transcript dataframe}
}
\value{
a long dataframe with allele counts by cell
}
\description{
Preprocess allele data
}
