% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_exp_fc}
\alias{process_exp_fc}
\title{Process single-cell gene expression data}
\usage{
process_exp_fc(count_mat_obs, lambdas_ref, gtf_transcript, verbose = TRUE)
}
\arguments{
\item{lambdas_ref}{expression values in reference profile}

\item{gtf_transcript}{transcript dataframe}

\item{count_mat}{observed gene count matrices}
}
\value{
logx+1 transformed normalized expression values for single cells in a long dataframe
}
\description{
Process single-cell gene expression data
}
