% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_exp_sc}
\alias{process_exp_sc}
\title{process single-cell gene expression data}
\usage{
process_exp_sc(
  count_mat,
  lambdas_ref,
  gtf_transcript,
  window = 101,
  verbose = T
)
}
\arguments{
\item{count_mat}{dgCMatrix Gene expression counts}

\item{lambdas_ref}{matrix Reference expression profiles}

\item{gtf_transcript}{dataframe Transcript gtf}
}
\value{
dataframe Logx+1 transformed normalized expression values for single cells
}
\description{
process single-cell gene expression data
}
\keyword{internal}
