% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_hca}
\alias{ref_hca}
\title{reference expression magnitudes from HCA}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 24756 rows and 12 columns.
}
\usage{
ref_hca
}
\description{
reference expression magnitudes from HCA
}
\keyword{datasets}
