% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{relevel_chrom}
\alias{relevel_chrom}
\title{Relevel chromosome column}
\usage{
relevel_chrom(df)
}
\arguments{
\item{df}{dataframe Dataframe with chromosome column}
}
\description{
Relevel chromosome column
}
\keyword{internal}
