% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{retest_bulks}
\alias{retest_bulks}
\title{retest consensus segments on pseudobulks}
\usage{
retest_bulks(bulks, segs_consensus, use_loh = FALSE, diploid_chroms = NULL)
}
\arguments{
\item{bulks}{pseudobulks dataframe}

\item{segs_consensus}{consensus segments dataframe}

\item{use_loh}{whether to use loh in the baseline}

\item{diploid_chroms}{user-provided diploid chromosomes}
}
\value{
pseudobulks dataframe
}
\description{
retest consensus segments on pseudobulks
}
