% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retest_cnv}
\alias{retest_cnv}
\title{retest CNVs in a pseudobulk}
\usage{
retest_cnv(bulk, exp_model = "lnpois", gamma = 20, allele_only = FALSE)
}
\arguments{
\item{bulk}{pesudobulk dataframe}

\item{exp_model}{model for gene expression - Poisson-Lognormal or Gamma-Poisson}

\item{gamma}{dispersion parameter for the Beta-Binomial allele model}

\item{allele_only}{whether to retest only using allele data}
}
\value{
a dataframe of segments with CNV posterior information
}
\description{
retest CNVs in a pseudobulk
}
