% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{return_missing_columns}
\alias{return_missing_columns}
\title{Check the format of a given file}
\usage{
return_missing_columns(file, expected_colnames = NULL)
}
\description{
Check the format of a given file
}
\keyword{internal}
