% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{run_group_hmms}
\alias{run_group_hmms}
\title{Run mutitple HMM}
\usage{
run_group_hmms(
  groups,
  count_mat,
  df,
  lambdas_ref,
  gtf_transcript,
  genetic_map,
  t,
  gamma = 20,
  min_depth = 0,
  ncores = NULL,
  exp_model = "lnpois",
  allele_only = FALSE,
  verbose = FALSE,
  debug = FALSE
)
}
\description{
Run mutitple HMM
}
