% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{run_hmm_inhom}
\alias{run_hmm_inhom}
\title{allele-only HMM}
\usage{
run_hmm_inhom(
  pAD,
  DP,
  p_s,
  t = 1e-05,
  theta_min = 0.08,
  gamma = 20,
  prior = NULL
)
}
\arguments{
\item{pAD}{integer vector Paternal allele counts}

\item{DP}{integer vector Total alelle counts}

\item{p_s}{numeric vector Phase switch probabilities}

\item{t}{numeric Transition probability between copy number states}

\item{theta_min}{numeric Minimum haplotype frequency deviation threshold}

\item{gamma}{numeric Overdispersion in the allele-specific expression}
}
\value{
character vector Decoded states
}
\description{
allele-only HMM
}
\keyword{internal}
