% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{run_hmm_mv_inhom}
\alias{run_hmm_mv_inhom}
\title{one theta level}
\usage{
run_hmm_mv_inhom(
  pAD,
  DP,
  p_s,
  Y_obs = 0,
  lambda_ref = 0,
  d_total = 0,
  theta_min = 0.08,
  theta_neu = 0,
  bal_cnv = TRUE,
  phi_neu = 1,
  phi_del = 2^(-0.25),
  phi_amp = 2^(0.25),
  phi_bamp = 2^(0.25),
  phi_bdel = 2^(-0.25),
  alpha = 1,
  beta = 1,
  mu = 0,
  sig = 1,
  exp_model = "gpois",
  t = 1e-05,
  gamma = 18,
  prior = NULL,
  exp_only = FALSE,
  allele_only = FALSE,
  classify_allele = FALSE,
  phasing = TRUE,
  debug = FALSE
)
}
\description{
one theta level
}
\keyword{internal}
