% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{run_hmm_mv_inhom2}
\alias{run_hmm_mv_inhom2}
\title{two theta level}
\usage{
run_hmm_mv_inhom2(
  pAD,
  DP,
  p_s,
  Y_obs,
  lambda_ref,
  d_total,
  theta_min = 0.08,
  theta_neu = 0.065,
  bal_cnv = TRUE,
  phi_neu = 1,
  phi_del = 2^(-0.25),
  phi_amp = 2^(0.25),
  phi_bamp = 2^(0.25),
  phi_bdel = 2^(-0.25),
  alpha = 1,
  beta = 1,
  mu = 0,
  sig = 1,
  exp_model = "lnpois",
  t = 1e-05,
  gamma = 20,
  prior = NULL,
  exp_only = FALSE,
  allele_only = FALSE,
  classify_allele = FALSE,
  phasing = TRUE,
  debug = FALSE
)
}
\description{
two theta level
}
\keyword{internal}
