% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_numbat}
\alias{run_numbat}
\title{Run workflow to decompose tumor subclones}
\usage{
run_numbat(
  count_mat,
  lambdas_ref,
  df_allele,
  gtf_transcript,
  genetic_map,
  out_dir = "./",
  max_iter = 2,
  t = 1e-05,
  gamma = 20,
  min_LLR = 50,
  alpha = 1e-04,
  eps = 1e-05,
  max_entropy = 0.6,
  init_k = 3,
  sample_size = 1e+05,
  min_cells = 10,
  max_cost = ncol(count_mat) * 0.3,
  min_depth = 0,
  common_diploid = TRUE,
  min_overlap = 0.45,
  ncores = 1,
  ncores_nni = 1,
  exp_model = "lnpois",
  verbose = TRUE,
  diploid_chroms = NULL,
  use_loh = NULL,
  exclude_normal = FALSE,
  skip_nj = FALSE,
  multi_allelic = FALSE,
  p_multi = 0.995,
  plot = TRUE
)
}
\arguments{
\item{count_mat}{raw count matrices where rownames are genes and column names are cells}

\item{lambdas_ref}{either a named vector with gene names as names and normalized expression as values, or a matrix where rownames are genes and columns are pseudobulk names}

\item{df_allele}{dataframe of allele counts per cell, produced by preprocess_allele}

\item{gtf_transcript}{gtf dataframe of transcripts}

\item{genetic_map}{a genetic map dataframe}

\item{out_dir}{output directory}

\item{max_iter}{maximum number of iterations to run the phyologeny optimization}

\item{t}{transition probability}

\item{gamma}{dispersion parameter for the Beta-Binomial allele model}

\item{min_LLR}{minimum LLR to filter CNVs by}

\item{eps}{convergence threshold for ML tree search}

\item{max_entropy}{entorpy threshold to filter CNVs by}

\item{init_k}{number of clusters in the initial clustering}

\item{sample_size}{subsampling size}

\item{min_cells}{minimum number of cells to run HMM on}

\item{max_cost}{maximum cost to simplify the phylogeny}

\item{common_diploid}{whether to find common diploid regions in a group of peusdobulks}

\item{ncores}{number of threads to use}

\item{diploid_chroms}{known diploid chromosomes}

\item{use_loh}{whether to include LOH regions in the expression baseline}

\item{exclude_normal}{whether to exclude normal cells in the tree construction}

\item{skip_nj}{whether to skip NJ tree construction and only use UPGMA}

\item{multi_allelic}{whether to call multi-allelic CNVs}

\item{min_dpeth}{minimum allele depth}
}
\value{
a status code
}
\description{
Run workflow to decompose tumor subclones
}
