% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_numbat_rdr}
\alias{run_numbat_rdr}
\title{Run Numbat RDR module}
\usage{
run_numbat_rdr(
  count_mat,
  lambdas_ref,
  genome = "hg38",
  out_dir = tempdir(),
  init_k = 3,
  ncores = 1,
  random_init = FALSE,
  verbose = TRUE,
  plot = TRUE
)
}
\arguments{
\item{count_mat}{dgCMatrix Raw count matrices where rownames are genes and column names are cells}

\item{lambdas_ref}{matrix Either a named vector with gene names as names and normalized expression as values, or a matrix where rownames are genes and columns are pseudobulk names}

\item{genome}{character Genome version (hg38, hg19, or mm10)}

\item{out_dir}{string Output directory}

\item{init_k}{integer Number of clusters in the initial clustering}

\item{ncores}{integer Number of threads to use}

\item{random_init}{logical Whether to initiate phylogney using a random tree (internal use only)}

\item{verbose}{logical Verbosity}

\item{plot}{logical Whether to plot results}
}
\value{
a status code
}
\description{
Run Numbat RDR module
}
