% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{score_mut}
\alias{score_mut}
\title{Score mutations on goodness of fit on the tree}
\usage{
score_mut(gtree, geno, p_min = 0.95)
}
\description{
Score mutations on goodness of fit on the tree
}
\keyword{internal}
