% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{score_tree}
\alias{score_tree}
\title{score a tree based on maximum likelihood}
\usage{
score_tree(tree, P, get_l_matrix = FALSE)
}
\arguments{
\item{tree}{phylo object}

\item{P}{genotype probability matrix}

\item{get_l_matrix}{whether to compute the whole likelihood matrix}
}
\description{
score a tree based on maximum likelihood
}
