% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segs_example}
\alias{segs_example}
\title{example CNV segments dataframe}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 27 rows and 30 columns.
}
\usage{
segs_example
}
\description{
example CNV segments dataframe
}
\keyword{datasets}
