% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smooth_segs}
\alias{smooth_segs}
\title{Smooth the segments after HMM decoding}
\usage{
smooth_segs(bulk, min_genes = 10)
}
\arguments{
\item{bulk}{a pseudobulk dataframe}

\item{min_genes}{minimum number of genes to call a segment}
}
\value{
a pseudobulk dataframe
}
\description{
Smooth the segments after HMM decoding
}
\keyword{internal}
