% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_prob_cm}
\alias{switch_prob_cm}
\title{predict phase switch probablity as a function of genetic distance}
\usage{
switch_prob_cm(d, lambda = 1, min_p = 1e-10)
}
\arguments{
\item{d}{genetic distance in cM}

\item{lambda}{rate of phase switch}

\item{min_p}{minimum phase switch probability}
}
\value{
phase switch probability
}
\description{
predict phase switch probablity as a function of genetic distance
}
\keyword{internal}
