% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t.test.pval}
\alias{t.test.pval}
\title{T-test wrapper, handles error for insufficient observations}
\usage{
\method{t.test}{pval}(x, y)
}
\description{
T-test wrapper, handles error for insufficient observations
}
\keyword{internal}
