% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t_test_pval}
\alias{t_test_pval}
\title{T-test wrapper, handles error for insufficient observations}
\usage{
t_test_pval(x, y)
}
\description{
T-test wrapper, handles error for insufficient observations
}
\keyword{internal}
