% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{test_multi_allelic}
\alias{test_multi_allelic}
\title{test for multi-allelic CNVs}
\usage{
test_multi_allelic(
  bulks,
  segs_consensus,
  use_loh = FALSE,
  LLR_min = 100,
  p_min = 0.995,
  diploid_chroms = NULL
)
}
\description{
test for multi-allelic CNVs
}
\keyword{internal}
