% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta_hat_roll}
\alias{theta_hat_roll}
\title{Rolling estimate of imbalance level theta}
\usage{
theta_hat_roll(major_count, minor_count, h)
}
\arguments{
\item{major_count}{vector of major allele count}

\item{minor_count}{vector of minor allele count}

\item{h}{window size}
}
\value{
rolling estimate of theta
}
\description{
Rolling estimate of imbalance level theta
}
\keyword{internal}
