% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta_hat_seg}
\alias{theta_hat_seg}
\title{Estimate of imbalance level theta in a segment}
\usage{
theta_hat_seg(major_count, minor_count)
}
\arguments{
\item{major_count}{vector of major allele count}

\item{minor_count}{vector of minor allele count}
}
\value{
estimate of theta
}
\description{
Estimate of imbalance level theta in a segment
}
\keyword{internal}
