% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{to_phylo}
\alias{to_phylo}
\title{Convert the phylogeny from tidygraph to igraph object}
\usage{
to_phylo(gtree)
}
\arguments{
\item{gtree}{tbl_graph The single-cell phylogeny}
}
\value{
phylo The single-cell phylogeny
}
\description{
Convert the phylogeny from tidygraph to igraph object
}
\keyword{internal}
