% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{transfer_links}
\alias{transfer_links}
\title{Annotate the direct upstream or downstream node on the edges}
\usage{
transfer_links(G)
}
\arguments{
\item{G}{igraph Mutation graph}
}
\value{
igraph Mutation graph
}
\description{
Annotate the direct upstream or downstream node on the edges
}
\keyword{internal}
