% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{upgma}
\alias{upgma}
\title{UPGMA and WPGMA clustering}
\usage{
upgma(D, method = "average", ...)
}
\arguments{
\item{D}{A distance matrix.}

\item{method}{The agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of "ward", "single", "complete", "average",
"mcquitty", "median" or "centroid". The default is "average".}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
UPGMA and WPGMA clustering
}
