% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{Viterbi.dthmm.mv.inhom.gpois}
\alias{Viterbi.dthmm.mv.inhom.gpois}
\title{Viterbi algorithm for joint HMM (gamma-poisson expression likelihood)}
\usage{
Viterbi.dthmm.mv.inhom.gpois(object, ...)
}
\description{
Viterbi algorithm for joint HMM (gamma-poisson expression likelihood)
}
\keyword{internal}
