% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{Viterbi.dthmm.mv.inhom.lnpois}
\alias{Viterbi.dthmm.mv.inhom.lnpois}
\title{Viterbi algorithm for joint HMM (PLN expression likelihood)}
\usage{
Viterbi.dthmm.mv.inhom.lnpois(object, ...)
}
\description{
Viterbi algorithm for joint HMM (PLN expression likelihood)
}
\keyword{internal}
