% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{viterbi_joint}
\alias{viterbi_joint}
\title{Viterbi algorithm for joint HMM (PLN expression likelihood)}
\usage{
viterbi_joint(object, ...)
}
\description{
Viterbi algorithm for joint HMM (PLN expression likelihood)
}
\keyword{internal}
