% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{viterbi_loh}
\alias{viterbi_loh}
\title{Viterbi for clonal LOH detection}
\usage{
viterbi_loh(hmm, ...)
}
\arguments{
\item{hmm}{HMM object; expect variables x (SNP count), snp_sig (snp rate standard deviation), 
pm (snp density for ref and loh states), pn (gene lengths),
d (total expression depth), y (expression count), lambda_star (reference expression rate), 
mu (global expression mean), sig (global expression standard deviation),
Pi (transition prob matrix), delta (prior for each state),
phi (expression fold change for each state)}
}
\description{
Viterbi for clonal LOH detection
}
\keyword{internal}
