% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_counts}
\alias{aggregate_counts}
\title{Utility function to make reference gene expression profiles}
\usage{
aggregate_counts(count_mat, cell_annot, verbose = T)
}
\arguments{
\item{count_mat}{dgCMatrix Gene expression counts}

\item{cell_annot}{dataframe Cell type annotation with columns "cell" and "cell_type"}

\item{verbose}{logical Verbosity}
}
\value{
matrix Reference gene expression levels
}
\description{
Utility function to make reference gene expression profiles
}
