% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_counts}
\alias{aggregate_counts}
\title{Utility function to make reference gene expression profiles}
\usage{
aggregate_counts(count_mat, annot, normalized = TRUE, verbose = TRUE)
}
\arguments{
\item{count_mat}{matrix/dgCMatrix Gene expression counts}

\item{annot}{dataframe Cell annotation with columns "cell" and "group"}

\item{normalized}{logical Whether to return normalized expression values}

\item{verbose}{logical Verbosity}
}
\value{
matrix Reference gene expression levels
}
\description{
Utility function to make reference gene expression profiles
}
\examples{
ref_custom = aggregate_counts(count_mat_ref, annot_ref, verbose = FALSE)
}
