% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analyze_bulk}
\alias{analyze_bulk}
\title{Call CNVs in a pseudobulk}
\usage{
analyze_bulk(
  bulk,
  t = 1e-05,
  gamma = 20,
  theta_min = 0.08,
  logphi_min = 0.25,
  lambda = 1,
  min_genes = 10,
  exp_only = FALSE,
  allele_only = FALSE,
  retest = TRUE,
  find_diploid = TRUE,
  diploid_chroms = NULL,
  classify_allele = FALSE,
  run_hmm = TRUE,
  bal_cnv = TRUE,
  prior = NULL,
  phasing = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{bulk}{dataframe Pesudobulk profile}

\item{t}{numeric Transition probability}

\item{gamma}{numeric Dispersion parameter for the Beta-Binomial allele model}

\item{theta_min}{numeric Minimum imbalance threshold}

\item{lambda}{numeric Phase switch rate}

\item{exp_only}{whether to run expression-only HMM}

\item{allele_only}{whether to run allele-only HMM}

\item{retest}{whether to retest CNVs after Viterbi decoding}

\item{find_diploid}{whether to run diploid region identification routine}

\item{diploid_chroms}{user-given chromsomes that are known to be in diploid state}
}
\value{
a dataframe of segments with CNV posterior information
}
\description{
Call CNVs in a pseudobulk
}
