% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annot_cm}
\alias{annot_cm}
\title{Annotate genetic distance between markers}
\usage{
annot_cm(bulk, genetic_map)
}
\arguments{
\item{bulk}{dataframe Pseudobulk profile}

\item{genetic_map}{dataframe Genetic map}
}
\value{
dataframe Annotated pseudobulk profile
}
\description{
Annotate genetic distance between markers
}
\keyword{internal}
