% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annot_consensus}
\alias{annot_consensus}
\title{Annotate a set of segments on a pseudobulk dataframe}
\usage{
annot_consensus(bulk, segs_consensus, join_mode = "inner")
}
\arguments{
\item{bulk}{dataframe Pseudobulk profile}

\item{segs_consensus}{datatframe Consensus segment dataframe}
}
\value{
dataframe Pseudobulk profile
}
\description{
Annotate a set of segments on a pseudobulk dataframe
}
\keyword{internal}
