% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annot_haplo_segs}
\alias{annot_haplo_segs}
\title{Annotate haplotype segments after HMM decoding}
\usage{
annot_haplo_segs(bulk)
}
\description{
Annotate haplotype segments after HMM decoding
}
\keyword{internal}
