% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annot_segs}
\alias{annot_segs}
\title{Annotate copy number segments after HMM decoding}
\usage{
annot_segs(bulk, var = "cnv_state")
}
\arguments{
\item{bulk}{dataframe Pseudobulk profile}
}
\value{
a pseudobulk dataframe
}
\description{
Annotate copy number segments after HMM decoding
}
\keyword{internal}
