% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_allele_LLR}
\alias{calc_allele_LLR}
\title{Calculate LLR for an allele HMM}
\usage{
calc_allele_LLR(pAD, DP, p_s, theta_mle, theta_0 = 0, gamma = 20)
}
\arguments{
\item{pAD}{numeric vector Phased allele depth}

\item{DP}{numeric vector Total allele depth}

\item{p_s}{numeric vector Phase switch probabilities}

\item{theta_mle}{numeric MLE of imbalance level theta (alternative hypothesis)}

\item{theta_0}{numeric Imbalance level in the null hypothesis}

\item{gamma}{numeric Dispersion parameter for the Beta-Binomial allele model}
}
\value{
numeric Log-likelihood ratio
}
\description{
Calculate LLR for an allele HMM
}
\keyword{internal}
