% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{calc_trans_mat}
\alias{calc_trans_mat}
\title{Calculate the transition matrix for joint HMM}
\usage{
calc_trans_mat(t, p_s, w, states_cn, states_phase)
}
\description{
Calculate the transition matrix for joint HMM
}
\keyword{internal}
