% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_ref_cor}
\alias{choose_ref_cor}
\title{choose beest reference for each cell based on correlation}
\usage{
choose_ref_cor(count_mat, lambdas_ref, gtf)
}
\arguments{
\item{count_mat}{dgCMatrix Gene expression counts}

\item{lambdas_ref}{matrix Reference expression profiles}

\item{gtf}{dataframe Transcript gtf}
}
\value{
named vector Best references for each cell
}
\description{
choose beest reference for each cell based on correlation
}
\keyword{internal}
