% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_alleles}
\alias{classify_alleles}
\title{Classify alleles using viterbi and forward-backward}
\usage{
classify_alleles(bulk)
}
\arguments{
\item{bulk}{dataframe Pesudobulk profile}
}
\value{
dataframe Pesudobulk profile
}
\description{
Classify alleles using viterbi and forward-backward
}
\keyword{internal}
