% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{cnv_heatmap}
\alias{cnv_heatmap}
\title{Plot CNV heatmap}
\usage{
cnv_heatmap(
  segs,
  var = "group",
  label_group = TRUE,
  legend = TRUE,
  exclude_gap = TRUE,
  genome = "hg38"
)
}
\arguments{
\item{segs}{dataframe Segments to plot. Need columns "seg_start", "seg_end", "cnv_state"}

\item{var}{character Column to facet by}

\item{label_group}{logical Label the groups}

\item{legend}{logical Display the legend}

\item{exclude_gap}{logical Whether to mark gap regions}

\item{genome}{character Genome build, either 'hg38' or 'hg19'}
}
\value{
ggplot Heatmap of CNVs along the genome
}
\description{
Plot CNV heatmap
}
\examples{
p = cnv_heatmap(segs_example)
}
