% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_bulk}
\alias{combine_bulk}
\title{Combine allele and expression pseudobulks}
\usage{
combine_bulk(allele_bulk, exp_bulk)
}
\arguments{
\item{allele_bulk}{dataframe Bulk allele profile}

\item{exp_bulk}{dataframe Bulk expression profile}
}
\value{
dataframe Pseudobulk allele and expression profile
}
\description{
Combine allele and expression pseudobulks
}
\keyword{internal}
