% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_posterior}
\alias{compute_posterior}
\title{Do bayesian averaging to get posteriors}
\usage{
compute_posterior(PL)
}
\arguments{
\item{PL}{dataframe Likelihoods and priors}
}
\value{
dataframe Posteriors
}
\description{
Do bayesian averaging to get posteriors
}
\keyword{internal}
