% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{exp_hclust}
\alias{exp_hclust}
\title{Run smoothed expression-based hclust}
\usage{
exp_hclust(
  count_mat,
  lambdas_ref,
  gtf_transcript,
  sc_refs = NULL,
  window = 101,
  k = 3,
  ncores = 1,
  verbose = T
)
}
\description{
Run smoothed expression-based hclust
}
\keyword{internal}
