% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{exp_hclust}
\alias{exp_hclust}
\title{Run smoothed expression-based hclust}
\usage{
exp_hclust(
  count_mat,
  lambdas_ref,
  gtf,
  sc_refs = NULL,
  window = 101,
  ncores = 1,
  verbose = TRUE
)
}
\arguments{
\item{count_mat}{dgCMatrix Gene counts}

\item{lambdas_ref}{matrix Reference expression profiles}

\item{gtf}{dataframe Transcript GTF}

\item{sc_refs}{named list Reference choices for single cells}

\item{window}{integer Sliding window size}

\item{ncores}{integer Number of cores}

\item{verbose}{logical Verbosity}
}
\description{
Run smoothed expression-based hclust
}
\keyword{internal}
