% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fill_neu_segs}
\alias{fill_neu_segs}
\title{Fill neutral regions into consensus segments}
\usage{
fill_neu_segs(segs_consensus, segs_neu)
}
\arguments{
\item{segs_consensus}{dataframe CNV segments from multiple samples}

\item{segs_neu}{dataframe Neutral segments}
}
\value{
dataframe Collections of neutral and aberrant segments with no gaps
}
\description{
Fill neutral regions into consensus segments
}
\keyword{internal}
