% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_genes}
\alias{filter_genes}
\title{filter for mutually expressed genes}
\usage{
filter_genes(count_mat, lambdas_ref, gtf, verbose = FALSE)
}
\arguments{
\item{count_mat}{dgCMatrix Gene expression counts}

\item{lambdas_ref}{named numeric vector A reference expression profile}

\item{gtf}{dataframe Transcript gtf}
}
\value{
vector Genes that are kept after filtering
}
\description{
filter for mutually expressed genes
}
\keyword{internal}
