% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_gpois}
\alias{fit_gpois}
\title{fit a Gamma-Poisson model by maximum likelihood}
\usage{
fit_gpois(Y_obs, lambda_ref, d)
}
\arguments{
\item{Y_obs}{numeric vector Gene expression counts}

\item{lambda_ref}{numeric vector Reference expression levels}

\item{d}{numeric Total library size}
}
\value{
numeric MLE of alpha and beta
}
\description{
fit a Gamma-Poisson model by maximum likelihood
}
\keyword{internal}
