% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_multi_ref}
\alias{fit_multi_ref}
\title{fit reference weights for a pseudobulk expression profile}
\usage{
fit_multi_ref(Y_obs, lambdas_ref, d, gtf, min_lambda = 2e-06, verbose = FALSE)
}
\arguments{
\item{Y_obs}{numeric vector Gene expression counts}

\item{lambdas_ref}{matrix Reference expression profiles}

\item{d}{numeric Total library size}

\item{gtf}{dataframe Transcript gtf}

\item{min_lambda}{numeric Expression level threshold to filter genes}
}
\value{
list Fitted reference weights and expression model parameters
}
\description{
fit reference weights for a pseudobulk expression profile
}
\keyword{internal}
