% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{forward_back_allele}
\alias{forward_back_allele}
\title{Forward-backward algorithm for allele HMM}
\usage{
forward_back_allele(hmm)
}
\arguments{
\item{hmm}{HMM object; expect variables x (allele depth), d (total depth),
logPi (log transition prob matrix), delta (prior for each state), 
alpha (alpha for each state), beta (beta for each state), 
states (states), p_s (phase switch probs)}
}
\description{
Forward-backward algorithm for allele HMM
}
\keyword{internal}
