% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{genotype}
\alias{genotype}
\title{Main function}
\usage{
genotype(label, samples, vcfs, outdir, het_only = FALSE, chr_prefix = TRUE)
}
\arguments{
\item{label}{character string}

\item{samples}{list of strings}

\item{vcfs}{vcfR objects}

\item{outdir}{output directory}

\item{het_only}{boolean (default=FALSE)}

\item{chr_prefix}{boolean (default=TRUE)}
}
\value{
a status code
}
\description{
Main function
}
