% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{genotype}
\alias{genotype}
\title{Genotyping main function}
\usage{
genotype(label, samples, vcfs, outdir, het_only = FALSE, chr_prefix = TRUE)
}
\arguments{
\item{label}{character Individual/sample label}

\item{samples}{vector Sample names}

\item{vcfs}{list of vcfR VCFs from cellsnp-lite pileup}

\item{outdir}{character Output directory}

\item{het_only}{logical Whether to only use heterozygous SNPs}

\item{chr_prefix}{logical Whether to add chr prefix}
}
\value{
integer Status code
}
\description{
Genotyping main function
}
\keyword{internal}
