% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_allele_bulk}
\alias{get_allele_bulk}
\title{Aggregate into pseudobulk alelle profile}
\usage{
get_allele_bulk(df_allele, nu = 1, min_depth = 0)
}
\arguments{
\item{df_allele}{dataframe Single-cell allele counts}

\item{nu}{numeric Phase switch rate}

\item{min_depth}{integer Minimum coverage to filter SNPs}
}
\value{
dataframe Pseudobulk allele profile
}
\description{
Aggregate into pseudobulk alelle profile
}
\keyword{internal}
