% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_allele_post}
\alias{get_allele_post}
\title{get CNV allele posteriors}
\usage{
get_allele_post(df_allele, haplotypes, segs_consensus)
}
\arguments{
\item{df_allele}{dataframe Allele counts}

\item{haplotypes}{dataframe Haplotype classification}

\item{segs_consensus}{dataframe Consensus CNV segments}
}
\value{
dataframe Allele posteriors
}
\description{
get CNV allele posteriors
}
\keyword{internal}
