% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bulk}
\alias{get_bulk}
\title{Aggregate into combined bulk expression and allele profile}
\usage{
get_bulk(
  count_mat,
  lambdas_ref,
  df_allele,
  gtf_transcript,
  genetic_map,
  min_depth = 0,
  lambda = 2,
  verbose = TRUE
)
}
\arguments{
\item{count_mat}{observed gene count matrices}

\item{lambdas_ref}{expression values in reference profile}

\item{df_allele}{allele dataframe}

\item{gtf_transcript}{transcript dataframe}
}
\value{
a dataframe of bulk gene expression and allele profile
}
\description{
Aggregate into combined bulk expression and allele profile
}
