% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bulk}
\alias{get_bulk}
\title{Aggregate into combined bulk expression and allele profile}
\usage{
get_bulk(
  count_mat,
  lambdas_ref,
  df_allele,
  gtf,
  genetic_map,
  min_depth = 0,
  lambda = 1,
  verbose = TRUE
)
}
\arguments{
\item{count_mat}{dgCMatrix Gene expression counts}

\item{lambdas_ref}{matrix Reference expression profiles}

\item{df_allele}{dataframe Single-cell allele counts}

\item{gtf}{dataframe Transcript gtf}

\item{genetic_map}{dataframe Genetic map}
}
\value{
dataframe Pseudobulk gene expression and allele profile
}
\description{
Aggregate into combined bulk expression and allele profile
}
