% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_exp_bulk}
\alias{get_exp_bulk}
\title{Aggregate into bulk expression profile}
\usage{
get_exp_bulk(count_mat, lambdas_ref, gtf_transcript, verbose = TRUE)
}
\arguments{
\item{count_mat}{dgCMatrix Gene expression counts}

\item{lambdas_ref}{matrix Reference expression profiles}

\item{gtf_transcript}{dataframe Transcript gtf}
}
\value{
dataframe Pseudobulk gene expression profile
}
\description{
Aggregate into bulk expression profile
}
\keyword{internal}
