% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_exp_likelihoods}
\alias{get_exp_likelihoods}
\title{get the single cell expression likelihoods}
\usage{
get_exp_likelihoods(
  exp_counts,
  diploid_chroms = NULL,
  use_loh = FALSE,
  depth_obs = NULL,
  mu = NULL,
  sigma = NULL
)
}
\arguments{
\item{exp_counts}{dataframe Single-cell expression counts (CHROM, seg, cnv_state, gene, Y_obs, lambda_ref)}

\item{diploid_chroms}{character vector Known diploid chromosomes}

\item{use_loh}{logical Whether to include CNLOH regions in baseline}
}
\value{
dataframe Single-cell CNV likelihood scores
}
\description{
get the single cell expression likelihoods
}
\keyword{internal}
