% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{get_exp_post}
\alias{get_exp_post}
\title{get CNV expression posteriors}
\usage{
get_exp_post(
  segs_consensus,
  count_mat,
  gtf_transcript,
  lambdas_ref = NULL,
  lambdas_fit = NULL,
  alpha = NULL,
  beta = NULL,
  ncores = 30,
  verbose = T,
  debug = F
)
}
\description{
get CNV expression posteriors
}
