% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_exp_sc}
\alias{get_exp_sc}
\title{get the single cell expression dataframe}
\usage{
get_exp_sc(segs_consensus, count_mat, gtf, segs_loh = NULL)
}
\arguments{
\item{segs_consensus}{dataframe Consensus segments}

\item{count_mat}{dgCMatrix gene expression count matrix}

\item{gtf}{dataframe Transcript gtf}
}
\value{
dataframe single cell expression counts annotated with segments
}
\description{
get the single cell expression dataframe
}
\keyword{internal}
