% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lambdas_bar}
\alias{get_lambdas_bar}
\title{Get average reference expressio profile based on single-cell ref choices}
\usage{
get_lambdas_bar(lambdas_ref, sc_refs, verbose = TRUE)
}
\arguments{
\item{lambdas_ref}{matrix Reference expression profiles}

\item{sc_refs}{vector Single-cell reference choices}

\item{verbose}{logical Print messages}
}
\description{
Get average reference expressio profile based on single-cell ref choices
}
\keyword{internal}
