% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_segs_consensus}
\alias{get_segs_consensus}
\title{Extract consensus CNV segments}
\usage{
get_segs_consensus(bulks, min_LLR = 5, min_overlap = 0.45, retest = TRUE)
}
\arguments{
\item{bulks}{dataframe Pseudobulks}

\item{min_LLR}{numeric LLR threshold to filter CNVs}

\item{min_overlap}{numeric Minimum overlap fraction to determine count two events as as overlapping}
}
\value{
dataframe Consensus segments
}
\description{
Extract consensus CNV segments
}
\keyword{internal}
