% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtf_hg19}
\alias{gtf_hg19}
\title{gene model (hg19)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 26841 rows and 5 columns.
}
\usage{
gtf_hg19
}
\description{
gene model (hg19)
}
\keyword{datasets}
