% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_bbinom}
\alias{l_bbinom}
\title{calculate joint likelihood of allele data}
\usage{
l_bbinom(AD, DP, alpha, beta)
}
\arguments{
\item{AD}{numeric vector Variant allele depth}

\item{DP}{numeric vector Total allele depth}

\item{alpha}{numeric Alpha parameter of Beta-Binomial distribution}

\item{beta}{numeric Beta parameter of Beta-Binomial distribution}
}
\value{
numeric Joint log likelihood
}
\description{
calculate joint likelihood of allele data
}
\keyword{internal}
