% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_gpois}
\alias{l_gpois}
\title{calculate joint likelihood of a gamma-poisson model}
\usage{
l_gpois(Y_obs, lambda_ref, d, alpha, beta, phi = 1)
}
\arguments{
\item{Y_obs}{numeric vector Gene expression counts}

\item{lambda_ref}{numeric vector Reference expression levels}

\item{d}{numeric Total library size}

\item{alpha}{numeric Alpha parameter of Gamma-Poisson distribution}

\item{beta}{numeric Beta parameter of Gamma-Poisson distribution}
}
\value{
numeric Joint log likelihood
}
\description{
calculate joint likelihood of a gamma-poisson model
}
\keyword{internal}
