% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_lnpois}
\alias{l_lnpois}
\title{ccalculate joint likelihood of a PLN model}
\usage{
l_lnpois(Y_obs, lambda_ref, d, mu, sig, phi = 1)
}
\arguments{
\item{Y_obs}{numeric vector Gene expression counts}

\item{lambda_ref}{numeric vector Reference expression levels}

\item{d}{numeric Total library size}
}
\value{
numeric Joint log likelihood
}
\description{
ccalculate joint likelihood of a PLN model
}
\keyword{internal}
