% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_group_bulks}
\alias{make_group_bulks}
\title{Make a group of pseudobulks}
\usage{
make_group_bulks(
  groups,
  count_mat,
  df_allele,
  lambdas_ref,
  gtf,
  genetic_map,
  min_depth = 0,
  ncores = NULL
)
}
\arguments{
\item{groups}{list Contains fields named "sample", "cells", "size", "members"}
}
\description{
Make a group of pseudobulks
}
\keyword{internal}
