% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{mark_tumor_lineage}
\alias{mark_tumor_lineage}
\title{Mark the tumor lineage of a phylogeny}
\usage{
mark_tumor_lineage(gtree)
}
\arguments{
\item{gtree}{tbl_graph Single-cell phylogeny}
}
\value{
tbl_graph Phylogeny annotated with tumor versus normal compartment
}
\description{
Mark the tumor lineage of a phylogeny
}
\keyword{internal}
