% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Numbat}
\alias{Numbat}
\title{Numbat R6 class}
\description{
Used to allow users to plot results
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{something something}

\item{\code{gtf}}{something something}

\item{\code{joint_post}}{something something}

\item{\code{exp_post}}{something something}

\item{\code{allele_post}}{something something}

\item{\code{bulk_subtrees}}{something something}

\item{\code{bulk_clones}}{something something}

\item{\code{segs_consensus}}{something something}

\item{\code{tree_post}}{something something}

\item{\code{mut_graph}}{something something}

\item{\code{gtree}}{something something}

\item{\code{clone_post}}{something something}

\item{\code{gexp_roll_wide}}{something something}

\item{\code{hc}}{something something}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Numbat$new()}}
\item \href{#method-plot_phylo_heatmap}{\code{Numbat$plot_phylo_heatmap()}}
\item \href{#method-plot_exp_roll}{\code{Numbat$plot_exp_roll()}}
\item \href{#method-plot_mut_history}{\code{Numbat$plot_mut_history()}}
\item \href{#method-plot_bulks}{\code{Numbat$plot_bulks()}}
\item \href{#method-clone}{\code{Numbat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize Numbat class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$new(out_dir, i = 2, gtf = gtf_hg38, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_dir}}{character string Output directory}

\item{\code{i}}{integer Get results from which iteration (either 1 or 2) (default=2)}

\item{\code{gtf}}{transcript gtf dataframe (default=gtf_hg38)}

\item{\code{verbose}}{boolean Whether to output verbose results (default=TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'numbat' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_phylo_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plot_phylo_heatmap}{}}}
\subsection{Method \code{plot_phylo_heatmap()}}{
Plot the single-cell CNV calls in a heatmap and the corresponding phylogeny
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$plot_phylo_heatmap(
  annot = NULL,
  geno_bar = FALSE,
  pal_clone = NULL,
  p_min = 0.9,
  tip_length = 1,
  branch_width = 0.2,
  line_width = 0.1,
  tree_height = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{annot}}{named list of cell annotation (default=NULL)}

\item{\code{geno_bar}}{where to show genotype annotation (default=NULL)}

\item{\code{pal_clone}}{clone color palette (default=NULL)}

\item{\code{p_min}}{minimum probability threshold to filter the calls (default=0.9)}

\item{\code{tip_length}}{tree tip length (default=1)}

\item{\code{branch_width}}{tree branch width (default=0.2)}

\item{\code{line_width}}{heatmap line width (default=0.1)}

\item{\code{tree_height}}{plotting height of the tree (default=1)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_exp_roll"></a>}}
\if{latex}{\out{\hypertarget{method-plot_exp_roll}{}}}
\subsection{Method \code{plot_exp_roll()}}{
Plot window-smoothed expression profiles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$plot_exp_roll(k = 3, n_sample = 300)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{integer Number of clusters (default=3)}

\item{\code{n_sample}}{integer Number of cells to subsample (default=300)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_mut_history"></a>}}
\if{latex}{\out{\hypertarget{method-plot_mut_history}{}}}
\subsection{Method \code{plot_mut_history()}}{
Plot the mutation history of the tumor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$plot_mut_history(horizontal = TRUE, label = TRUE, pal_clone = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{horizontal}}{horizontal layout or vertical}

\item{\code{label}}{whether to label the mutations on edges}

\item{\code{pal_clone}}{clone color palette}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_bulks"></a>}}
\if{latex}{\out{\hypertarget{method-plot_bulks}{}}}
\subsection{Method \code{plot_bulks()}}{
Plot the bulk CNV profiles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$plot_bulks(
  what = "clones",
  min_depth = 8,
  phi_mle = TRUE,
  ncol = 1,
  legend = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{whether to visualize clones or subtrees}

\item{\code{min_depth}}{minimum allele coverage to filter SNPs}

\item{\code{phi_mle}}{whether to plot expression fold change estimates}

\item{\code{ncol}}{number of columns in the plot panel}

\item{\code{legend}}{whether to display CNV state legend}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Numbat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
