% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{perform_nni}
\alias{perform_nni}
\title{Maximum likelihood tree search via NNI}
\usage{
perform_nni(
  tree_init,
  P,
  max_iter = 100,
  eps = 0.01,
  ncores = 1,
  verbose = TRUE
)
}
\arguments{
\item{tree_init}{phylo Intial tree}

\item{P}{matrix Genotype probability matrix}

\item{max_iter}{integer Maximum number of iterations}

\item{eps}{numeric Tolerance threshold in likelihood difference for stopping}

\item{ncores}{integer Number of cores to use}

\item{verbose}{logical Verbosity}
}
\value{
multiPhylo List of trees corresponding to the rearrangement steps
}
\description{
Maximum likelihood tree search via NNI
}
