% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phi_hat_roll}
\alias{phi_hat_roll}
\title{Rolling estimate of expression fold change phi}
\usage{
phi_hat_roll(Y_obs, lambda_ref, d_obs, mu, sig, h)
}
\description{
Rolling estimate of expression fold change phi
}
\keyword{internal}
