% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phi_hat_seg}
\alias{phi_hat_seg}
\title{Estimate of expression fold change phi in a segment}
\usage{
phi_hat_seg(Y_obs, lambda_ref, d, mu, sig)
}
\description{
Estimate of expression fold change phi in a segment
}
\keyword{internal}
