% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_consensus}
\alias{plot_consensus}
\title{Plot consensus CNVs}
\usage{
plot_consensus(segs)
}
\arguments{
\item{segs}{dataframe Consensus segments}
}
\value{
ggplot object
}
\description{
Plot consensus CNVs
}
\examples{
p = plot_consensus(segs_example)
}
