% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_mut_history}
\alias{plot_mut_history}
\title{Plot mutational history}
\usage{
plot_mut_history(
  G,
  clone_post = NULL,
  edge_label_size = 4,
  node_label_size = 6,
  node_size = 10,
  arrow_size = 2,
  show_clone_size = TRUE,
  show_distance = TRUE,
  legend = TRUE,
  edge_label = TRUE,
  node_label = TRUE,
  horizontal = TRUE,
  pal = NULL
)
}
\arguments{
\item{G}{igraph Mutation history graph}

\item{clone_post}{dataframe Clone assignment posteriors}

\item{edge_label_size}{numeric Size of edge label}

\item{node_label_size}{numeric Size of node label}

\item{node_size}{numeric Size of nodes}

\item{arrow_size}{numeric Size of arrows}

\item{show_clone_size}{logical Whether to show clone size}

\item{show_distance}{logical Whether to show evolutionary distance between clones}

\item{legend}{logical Whether to show legend}

\item{edge_label}{logical Whether to label edges}

\item{node_label}{logical Whether to label nodes}

\item{horizontal}{logical Whether to use horizontal layout}

\item{pal}{named vector Node colors}
}
\value{
ggplot object
}
\description{
Plot mutational history
}
\examples{
p = plot_mut_history(mut_graph_example)
}
