% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_psbulk}
\alias{plot_psbulk}
\title{plot a pseudobulk HMM profile}
\usage{
plot_psbulk(
  bulk,
  use_pos = FALSE,
  allele_only = FALSE,
  min_LLR = 10,
  min_depth = 8,
  exp_limit = 2,
  phi_mle = TRUE,
  theta_roll = FALSE,
  dot_size = 0.8,
  dot_alpha = 0.5,
  legend = TRUE
)
}
\arguments{
\item{bulk}{pseudobulk dataframe}

\item{use_pos}{use marker position instead of index as x coordinate}

\item{allele_only}{only plot alleles}

\item{min_LLR}{LLR threshold for event filtering}

\item{min_depth}{minimum coverage depth for a SNP to be plotted}

\item{exp_limit}{expression logFC axis limit}

\item{phi_mle}{whether to plot estimates of segmental expression fold change}

\item{theta_roll}{whether to plot rolling estimates of allele imbalance}

\item{dot_size}{size of marker dots}

\item{dot_alpha}{transparency of the marker dots}

\item{legend}{whether to show legend}
}
\value{
a ggplot object
}
\description{
plot a pseudobulk HMM profile
}
