% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_psbulk}
\alias{plot_psbulk}
\title{plot a pseudobulk HMM profile}
\usage{
plot_psbulk(
  bulk,
  use_pos = FALSE,
  allele_only = FALSE,
  min_LLR = 0,
  min_depth = 8,
  exp_limit = 2,
  phi_mle = TRUE,
  theta_roll = FALSE,
  dot_size = 0.8,
  dot_alpha = 0.5,
  legend = TRUE
)
}
\arguments{
\item{bulk}{dataframe Pseudobulk profile}

\item{use_pos}{logical Use marker position instead of index as x coordinate}

\item{allele_only}{logical Only plot alleles}

\item{min_LLR}{numeric LLR threshold for event filtering}

\item{min_depth}{numeric Minimum coverage depth for a SNP to be plotted}

\item{exp_limit}{numeric Expression logFC axis limit}

\item{phi_mle}{logical Whether to plot estimates of segmental expression fold change}

\item{theta_roll}{logical Whether to plot rolling estimates of allele imbalance}

\item{dot_size}{numeric Size of marker dots}

\item{dot_alpha}{numeric Transparency of the marker dots}

\item{legend}{logical Whether to show legend}
}
\value{
ggplot Plot of pseudobulk HMM profile
}
\description{
plot a pseudobulk HMM profile
}
