% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pre_likelihood_hmm}
\alias{pre_likelihood_hmm}
\title{Input to likelihood_allele() for unit tests
Based on example using bulk = get_bulk(count_mat = count_mat_ATC2, df_allele = df_allele_ATC2, lambdas_ref = ref_hca,gtf_transcript = gtf_hg38,genetic_map = genetic_map_hg38)}
\format{
An object of class \code{dthmm.inhom} of length 8.
}
\usage{
pre_likelihood_hmm
}
\description{
Input to likelihood_allele() for unit tests
Based on example using bulk = get_bulk(count_mat = count_mat_ATC2, df_allele = df_allele_ATC2, lambdas_ref = ref_hca,gtf_transcript = gtf_hg38,genetic_map = genetic_map_hg38)
}
\keyword{datasets}
