% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pre_likelihood_hmm}
\alias{pre_likelihood_hmm}
\title{HMM object for unit tests}
\format{
An object of class \code{list} of length 10.
}
\usage{
pre_likelihood_hmm
}
\description{
HMM object for unit tests
}
\keyword{datasets}
