% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{preprocess_allele}
\alias{preprocess_allele}
\title{Preprocess allele data}
\usage{
preprocess_allele(sample, vcf_pu, vcf_phased, AD, DP, barcodes, gtf, gmap)
}
\arguments{
\item{sample}{character Sample label}

\item{vcf_pu}{dataframe Pileup VCF from cell-snp-lite}

\item{vcf_phased}{dataframe Phased VCF from eagle2}

\item{AD}{dgTMatrix Alt allele depth matrix from pileup}

\item{DP}{dgTMatrix Total allele depth matrix from pileup}

\item{barcodes}{vector List of barcodes from pileup}

\item{gtf}{dataframe Transcript GTF}

\item{gmap}{dataframe Genetic map}
}
\value{
dataframe Allele counts by cell
}
\description{
Preprocess allele data
}
\keyword{internal}
