% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_exp_sc}
\alias{process_exp_sc}
\title{process single-cell gene expression data}
\usage{
process_exp_sc(
  count_mat,
  lambdas_ref,
  gtf_transcript,
  window = 101,
  verbose = T
)
}
\arguments{
\item{count_mat}{observed gene count matrices}

\item{lambdas_ref}{expression values in reference profile}

\item{gtf_transcript}{transcript dataframe}
}
\value{
logx+1 transformed normalized expression values for single cells in a long dataframe
}
\description{
process single-cell gene expression data
}
\keyword{internal}
