% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_hca_counts}
\alias{ref_hca_counts}
\title{reference expression counts from HCA}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 24857 rows and 12 columns.
}
\usage{
ref_hca_counts
}
\description{
reference expression counts from HCA
}
\keyword{datasets}
