% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{resolve_cnvs}
\alias{resolve_cnvs}
\title{Get unique CNVs from set of segments}
\usage{
resolve_cnvs(segs_all, min_overlap = 0.5, debug = FALSE)
}
\arguments{
\item{segs_all}{dataframe CNV segments from multiple samples}

\item{min_overlap}{numeric scalar Minimum overlap fraction to determine count two events as as overlapping}
}
\value{
dataframe Consensus CNV segments
}
\description{
Get unique CNVs from set of segments
}
\keyword{internal}
