% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{retest_bulks}
\alias{retest_bulks}
\title{retest consensus segments on pseudobulks}
\usage{
retest_bulks(
  bulks,
  segs_consensus = NULL,
  t = 1e-05,
  min_genes = 10,
  gamma = 20,
  nu = 1,
  use_loh = FALSE,
  diploid_chroms = NULL,
  ncores = 1,
  exclude_neu = TRUE,
  min_LLR = 5
)
}
\arguments{
\item{bulks}{dataframe Pseudobulk profiles}

\item{segs_consensus}{dataframe Consensus segments}

\item{use_loh}{logical Whether to use loh in the baseline}

\item{diploid_chroms}{vector User-provided diploid chromosomes}
}
\value{
dataframe Retested pseudobulks
}
\description{
retest consensus segments on pseudobulks
}
\keyword{internal}
