% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retest_cnv}
\alias{retest_cnv}
\title{retest CNVs in a pseudobulk}
\usage{
retest_cnv(
  bulk,
  theta_min = 0.08,
  logphi_min = 0.25,
  gamma = 20,
  allele_only = FALSE,
  exclude_neu = TRUE
)
}
\arguments{
\item{bulk}{pesudobulk dataframe}

\item{gamma}{numeric Dispersion parameter for the Beta-Binomial allele model}

\item{allele_only}{whether to retest only using allele data}
}
\value{
a dataframe of segments with CNV posterior information
}
\description{
retest CNVs in a pseudobulk
}
\keyword{internal}
