% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{run_group_hmms}
\alias{run_group_hmms}
\title{Run mutitple HMMs}
\usage{
run_group_hmms(
  bulks,
  t = 1e-04,
  gamma = 20,
  theta_min = 0.08,
  exp_model = "lnpois",
  allele_only = FALSE,
  retest = TRUE,
  ncores = NULL,
  verbose = FALSE,
  debug = FALSE
)
}
\description{
Run mutitple HMMs
}
