% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_group_hmms}
\alias{run_group_hmms}
\title{Run mutitple HMMs}
\usage{
run_group_hmms(
  bulks,
  t = 1e-04,
  gamma = 20,
  exp_model = "lnpois",
  alpha = 1e-04,
  min_genes = 10,
  common_diploid = TRUE,
  diploid_chroms = NULL,
  allele_only = FALSE,
  retest = TRUE,
  run_hmm = TRUE,
  ncores = NULL,
  verbose = FALSE,
  debug = FALSE
)
}
\description{
Run mutitple HMMs
}
\keyword{internal}
