% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_numbat}
\alias{run_numbat}
\title{Run workflow to decompose tumor subclones}
\usage{
run_numbat(
  count_mat,
  lambdas_ref,
  df_allele,
  gtf_transcript,
  genetic_map,
  out_dir = "./",
  max_iter = 2,
  max_nni = 100,
  t = 1e-05,
  gamma = 20,
  min_LLR = 50,
  alpha = 1e-04,
  eps = 1e-05,
  max_entropy = 0.6,
  init_k = 3,
  min_cells = 10,
  max_cost = ncol(count_mat) * 0.3,
  min_depth = 0,
  common_diploid = TRUE,
  min_overlap = 0.45,
  ncores = 1,
  ncores_nni = ncores,
  exp_model = "lnpois",
  verbose = TRUE,
  diploid_chroms = NULL,
  use_loh = NULL,
  skip_nj = FALSE,
  multi_allelic = FALSE,
  p_multi = 0.995,
  plot = TRUE
)
}
\arguments{
\item{count_mat}{dgCMatrix Raw count matrices where rownames are genes and column names are cells}

\item{lambdas_ref}{matrix Either a named vector with gene names as names and normalized expression as values, or a matrix where rownames are genes and columns are pseudobulk names}

\item{df_allele}{dataframe Allele counts per cell, produced by preprocess_allele}

\item{gtf_transcript}{dataframe GTF of transcripts}

\item{genetic_map}{dataframe A genetic map dataframe}

\item{out_dir}{string Output directory}

\item{max_iter}{integer Maximum number of iterations to run the phyologeny optimization}

\item{t}{numeric Transition probability}

\item{gamma}{numeric Dispersion parameter for the Beta-Binomial allele model}

\item{min_LLR}{numeric Minimum LLR to filter CNVs}

\item{eps}{numeric Convergence threshold for ML tree search}

\item{max_entropy}{numeric Entropy threshold to filter CNVs}

\item{init_k}{integer Number of clusters in the initial clustering}

\item{min_cells}{integer Minimum number of cells to run HMM on}

\item{max_cost}{numeric Maximum cost to simplify the phylogeny}

\item{common_diploid}{logical Whether to find common diploid regions in a group of peusdobulks}

\item{ncores}{integer Number of threads to use}

\item{diploid_chroms}{vector Known diploid chromosomes}

\item{use_loh}{logical Whether to include LOH regions in the expression baseline}

\item{skip_nj}{logical Whether to skip NJ tree construction and only use UPGMA}

\item{multi_allelic}{logical Whether to call multi-allelic CNVs}

\item{min_dpeth}{integer Minimum allele depth}
}
\value{
a status code
}
\description{
Run workflow to decompose tumor subclones
}
