% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smooth_expression}
\alias{smooth_expression}
\title{filtering, normalization and capping}
\usage{
smooth_expression(count_mat, lambdas_ref, gtf, window = 101, verbose = FALSE)
}
\arguments{
\item{count_mat}{dgCMatrix Gene expression counts}

\item{lambdas_ref}{matrix Reference expression profiles}

\item{gtf}{dataframe Transcript gtf}
}
\value{
dataframe Log(x+1) transformed normalized expression values for single cells
}
\description{
filtering, normalization and capping
}
\keyword{internal}
