% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_prob_cm}
\alias{switch_prob_cm}
\title{predict phase switch probablity as a function of genetic distance}
\usage{
switch_prob_cm(d, nu = 1, min_p = 1e-10)
}
\arguments{
\item{d}{numeric vector Genetic distance in cM}

\item{nu}{numeric Phase switch rate}

\item{min_p}{numeric Minimum phase switch probability}
}
\value{
numeric vector Phase switch probability
}
\description{
predict phase switch probablity as a function of genetic distance
}
\keyword{internal}
