% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{test_multi_allelic}
\alias{test_multi_allelic}
\title{test for multi-allelic CNVs}
\usage{
test_multi_allelic(bulks, segs_consensus, min_LLR = 5, p_min = 0.999)
}
\arguments{
\item{bulks}{dataframe Pseudobulk profiles}

\item{segs_consensus}{dataframe Consensus segments}

\item{min_LLR}{numeric CNV LLR threshold to filter events}

\item{p_min}{numeric Probability threshold to call multi-allelic events}
}
\value{
dataframe Consensus segments annotated with multi-allelic events
}
\description{
test for multi-allelic CNVs
}
\keyword{internal}
